% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mb.R
\name{add_mb}
\alias{add_mb}
\title{Add Mb}
\usage{
add_mb(dat, pos_col = "POS")
}
\arguments{
\item{dat}{Data.}

\item{pos_col}{Genomic position column name.}
}
\description{
Add column containing genomic coordinate in units of megabasepairs ("Mb").
}
\keyword{internal}
