% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_missense.R
\name{annotate_missense}
\alias{annotate_missense}
\title{Annotate any missense variants}
\usage{
annotate_missense(merged_DT, snp_filter = "Support>0")
}
\arguments{
\item{merged_DT}{Merged fine-mapping results data from
\link[echolocatoR]{finemap_loci}.}

\item{snp_filter}{Row-wise filter to apply to \code{merged_DT} filter 
(provided as a string).}
}
\description{
Annotate any missense variants with \link[biomaRt]{getBM}.
}
\examples{
\dontrun{
merged_DT <- echodata::get_Nalls2019_merged()
annotated_DT <- annotate_missense(merged_DT = merged_DT)
}
}
\seealso{
Other annotate: 
\code{\link{biomart_snp_info}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{haplor_epigenetics_enrichment}()},
\code{\link{haplor_epigenetics_summary}()},
\code{\link{haplor_haploreg}()},
\code{\link{haplor_regulomedb}()},
\code{\link{plot_missense}()},
\code{\link{snps_by_mutation_type}()}
}
\concept{annotate}
