% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_file_name.R
\name{annotation_file_name}
\alias{annotation_file_name}
\title{Annotation file name}
\usage{
annotation_file_name(locus_dir, lib_name, suffix = ".rds")
}
\arguments{
\item{locus_dir}{Locus-specific directory.}

\item{lib_name}{Annotation library name.}

\item{suffix}{File suffix.}
}
\description{
Construct an annotation-specific file name.
}
\examples{
tmp <- tempdir()
annot_file <- annotation_file_name(locus_dir=tmp,
                                   lib_name="test")
}
