% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_geneInfo.R
\name{biomart_geneInfo}
\alias{biomart_geneInfo}
\title{Get gene info using Biomart}
\source{
\href{https://bioconductor.org/packages/release/bioc/html/biomaRt.html}{
biomaRt}
}
\usage{
biomart_geneInfo(geneList, reference_genome = "grch37")
}
\description{
Get gene info using Biomart
}
\examples{
\dontrun{
gene_info <- biomart_geneInfo(c("PTK2B", "CLU", "APOE"))
}
}
\seealso{
Other annotate: 
\code{\link{annotate_missense}()},
\code{\link{biomart_snp_info}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{haplor_epigenetics_enrichment}()},
\code{\link{haplor_epigenetics_summary}()},
\code{\link{haplor_haploreg}()},
\code{\link{haplor_regulomedb}()},
\code{\link{plot_missense}()},
\code{\link{snps_by_mutation_type}()}
}
\concept{annotate}
\keyword{internal}
