% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_snp_info.R
\name{biomart_snp_info}
\alias{biomart_snp_info}
\title{Download SNP-wise annotations from Biomart}
\source{
\href{https://bioconductor.org/packages/release/bioc/html/biomaRt.html}{
biomaRt}
}
\usage{
biomart_snp_info(
  snp_list,
  reference_genome = "grch37",
  attributes = c("refsnp_id", "allele", "chr_name", "chrom_start", "chrom_end",
    "chrom_strand", "ensembl_gene_stable_id", "consequence_type_tv",
    "polyphen_prediction", "polyphen_score", "sift_prediction", "sift_score",
    "reg_consequence_types", "validated"),
  verbose = TRUE
)
}
\description{
Download SNP-wise annotations from Biomart
}
\seealso{
Other annotate: 
\code{\link{annotate_missense}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{haplor_epigenetics_enrichment}()},
\code{\link{haplor_epigenetics_summary}()},
\code{\link{haplor_haploreg}()},
\code{\link{haplor_regulomedb}()},
\code{\link{plot_missense}()},
\code{\link{snps_by_mutation_type}()}
}
\concept{annotate}
\keyword{internal}
