% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_snps_to_geneInfo.R
\name{biomart_snps_to_geneInfo}
\alias{biomart_snps_to_geneInfo}
\title{Identify which genes SNPs belong to using Biomart}
\source{
\href{https://bioconductor.org/packages/release/bioc/html/biomaRt.html}{
biomaRt}
}
\usage{
biomart_snps_to_geneInfo(snp_list, reference_genome = "grch37")
}
\description{
Identify which genes SNPs belong to using Biomart
}
\examples{
\dontrun{
# biomart_snps_to_geneInfo(c("rs114360492"))
}
}
\seealso{
Other annotate: 
\code{\link{annotate_missense}()},
\code{\link{biomart_geneInfo}()},
\code{\link{biomart_snp_info}()},
\code{\link{haplor_epigenetics_enrichment}()},
\code{\link{haplor_epigenetics_summary}()},
\code{\link{haplor_haploreg}()},
\code{\link{haplor_regulomedb}()},
\code{\link{plot_missense}()},
\code{\link{snps_by_mutation_type}()}
}
\concept{annotate}
\keyword{internal}
