% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_peaks.R
\name{call_peaks}
\alias{call_peaks}
\title{Call peaks}
\usage{
call_peaks(
  bedgraph_path,
  cutoff = NULL,
  outdir = tempdir(),
  outputfile = "bdg_peakcall.bed",
  return_path = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{bedgraph_path}{Path to bedGraph file.}

\item{cutoff}{Cutoff depends on which method you used for score track.
If the file contains pvalue scores from MACS3, score 5 means pvalue 1e-5.
If \code{NULL}, a reasonable \code{cutoff} value will be inferred 
through a \code{cutoff_analysis}.}

\item{outdir}{Directory to store \code{cutoff_analysis} 
report and peak file in.}

\item{outputfile}{Name of the peak output file (stored in BED format).}

\item{return_path}{Whether to return the path to the saved peak file, 
or the peak data itself as a \link[GenomicRanges]{GRanges} object.}

\item{verbose}{Print messages.}
}
\description{
Call peaks from a bedGraph file using \link[MACSr]{bdgpeakcall}, 
which is a wrapper for the \emph{MACS3} command line tool.
Can automatically infer a reasonable \code{cutoff} threshold as well.
}
\examples{

}
