% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_plots.R
\name{convert_plots}
\alias{convert_plots}
\title{Convert plots to various formats}
\usage{
convert_plots(
  plot_list,
  return_as = c("ggplot", "ggbio", "patchwork", "Tracks", NULL),
  x_limits = NULL,
  tag_levels = letters,
  params_list = NULL,
  verbose = TRUE
)
}
\arguments{
\item{plot_list}{A plot, or a list of plots belonging
 to one of the following classes:
\itemize{
\item{\link[ggplot2]{ggplot}}
\item{\link[ggbio]{ggbio}}
\item{\link[ggbio]{tracks}}
}}

\item{return_as}{Plot class to convert \code{plot_list} to:
\itemize{
\item{"ggplot"}{\link[ggplot2]{ggplot}}
\item{"ggbio"}{\link[ggbio]{ggbio}}
\item{"patchwork"}{\pkg{patchwork}}
\item{"Tracks"}{\link[ggbio]{tracks}}
\item{NULL}{Return original object.}
}}

\item{x_limits}{x-axis limits to be applied to all plots
(useful when trying to keep a common coordinate system).}

\item{tag_levels}{A character vector defining the enumeration format to use
at each level. Possible values are \code{'a'} for lowercase letters, \code{'A'} for
uppercase letters, \code{'1'} for numbers, \code{'i'} for lowercase Roman numerals, and
\code{'I'} for uppercase Roman numerals. It can also be a list containing
character vectors defining arbitrary tag sequences. If any element in the
list is a scalar and one of \code{'a'}, \code{'A'}, \code{'1'}, \verb{'i}, or \code{'I'}, this level
will be expanded to the expected sequence.}

\item{params_list}{A list of parameters to pass to \link[ggbio]{tracks}.}

\item{verbose}{Print messages.}
}
\description{
Takes a plot or just of plots and converts to various formats.
}
\examples{
#### Create example plot_list ####
gg <- ggplot2::ggplot(mtcars, ggplot2::aes(mpg, cyl)) + ggplot2::geom_point()
plot_list <- lapply(1:3, function(x) gg)

##### As ggplot #####
return_out1 <- convert_plots(plot_list = plot_list, return_as = "ggplot")
##### As ggbio #####
return_out2 <- convert_plots(plot_list = plot_list, return_as = "ggbio")
##### As patchwork #####
return_out3 <- convert_plots(plot_list = plot_list, return_as = "patchwork")
##### As Tracks #####
return_out4 <- convert_plots(plot_list = plot_list, return_as = "Tracks")
##### From Tracks #####
return_out5 <- echoannot::convert_plots(plot_list = return_out4)
}
