% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORCES2020_get_ATAC_peak_overlap.R
\name{CORCES2020_get_ATAC_peak_overlap}
\alias{CORCES2020_get_ATAC_peak_overlap}
\title{Get overlap between datatable of SNPs and scATAC peaks}
\source{
\url{https://doi.org/10.1038/s41588-020-00721-x}
}
\usage{
CORCES2020_get_ATAC_peak_overlap(
  query_dat,
  FDR_filter = NULL,
  add_cicero = TRUE,
  cell_type_specific = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{query_dat}{Genomic summary statistics data to query with.}

\item{FDR_filter}{Correct p-value threshold.}

\item{add_cicero}{Whether to include 
\href{https://www.bioconductor.org/packages/release/bioc/html/cicero.html}{
cicero} results as well.}

\item{cell_type_specific}{Whether to use bulk or cell-type-specific data.}

\item{verbose}{Print messages.}
}
\description{
Can optionally add \code{Cicero} coaccessibility scores,
which are also derived from scATAC-seq data.
}
\examples{
query_dat <- echodata::BST1[1:100,]
gr.hits <- echoannot::CORCES2020_get_ATAC_peak_overlap(query_dat = query_dat)
}
\seealso{
Other CORCES2020: 
\code{\link{CORCES2020_get_hichip_fithichip_overlap}()},
\code{\link{CORCES2020_prepare_bulkATAC_peak_overlap}()},
\code{\link{CORCES2020_prepare_scATAC_peak_overlap}()},
\code{\link{get_CORCES2020_bulkATACseq_peaks}()},
\code{\link{get_CORCES2020_cicero_coaccessibility}()},
\code{\link{get_CORCES2020_hichip_fithichip_loop_calls}()},
\code{\link{get_CORCES2020_scATACseq_celltype_peaks}()},
\code{\link{get_CORCES2020_scATACseq_peaks}()}
}
\concept{CORCES2020}
