% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORCES2020_get_hichip_fithichip_overlap.R
\name{CORCES2020_get_hichip_fithichip_overlap}
\alias{CORCES2020_get_hichip_fithichip_overlap}
\title{Get overlap between data table of SNPs and HiChIP_FitHiChIP
coaccessibility anchors}
\source{
\url{https://doi.org/10.1038/s41588-020-00721-x}
}
\usage{
CORCES2020_get_hichip_fithichip_overlap(query_dat, verbose = TRUE)
}
\arguments{
\item{query_dat}{Fine-mapping results.}

\item{verbose}{Print messages.}
}
\description{
Anchors are the genomic regions that have evidence of being
functionally connected to one another (coaccessible),
 e.g. enhancer-promoter interactions.
}
\seealso{
Other CORCES2020: 
\code{\link{CORCES2020_get_ATAC_peak_overlap}()},
\code{\link{CORCES2020_prepare_bulkATAC_peak_overlap}()},
\code{\link{CORCES2020_prepare_scATAC_peak_overlap}()},
\code{\link{get_CORCES2020_bulkATACseq_peaks}()},
\code{\link{get_CORCES2020_cicero_coaccessibility}()},
\code{\link{get_CORCES2020_hichip_fithichip_loop_calls}()},
\code{\link{get_CORCES2020_scATACseq_celltype_peaks}()},
\code{\link{get_CORCES2020_scATACseq_peaks}()}
}
\concept{CORCES2020}
