% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORCES2020_prepare_bulkATAC_peak_overlap.R
\name{CORCES2020_prepare_bulkATAC_peak_overlap}
\alias{CORCES2020_prepare_bulkATAC_peak_overlap}
\title{Prepare data to plot overlap between datatable of SNPs and
cell-type-specific epigenomic peaks and coaccessibility data.}
\source{
\url{https://doi.org/10.1038/s41588-020-00721-x}
}
\usage{
CORCES2020_prepare_bulkATAC_peak_overlap(
  merged_DT,
  FDR_filter = NULL,
  snp_filter = "Consensus_SNP==TRUE",
  add_HiChIP_FitHiChIP = TRUE,
  annotate_genes = FALSE,
  return_counts = TRUE,
  verbose = TRUE
)
}
\description{
Prepare data to plot overlap between datatable of SNPs and
cell-type-specific epigenomic peaks and coaccessibility data.
}
\examples{
\dontrun{
dat_melt <- echoannot:::CORCES2020_prepare_bulkATAC_peak_overlap(
    merged_DT = echodata::get_Nalls2019_merged()
)
}
}
\seealso{
Other CORCES2020: 
\code{\link{CORCES2020_get_ATAC_peak_overlap}()},
\code{\link{CORCES2020_get_hichip_fithichip_overlap}()},
\code{\link{CORCES2020_prepare_scATAC_peak_overlap}()},
\code{\link{get_CORCES2020_bulkATACseq_peaks}()},
\code{\link{get_CORCES2020_cicero_coaccessibility}()},
\code{\link{get_CORCES2020_hichip_fithichip_loop_calls}()},
\code{\link{get_CORCES2020_scATACseq_celltype_peaks}()},
\code{\link{get_CORCES2020_scATACseq_peaks}()}
}
\concept{CORCES2020}
\keyword{internal}
