% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORCES_2020.get_ATAC_peak_overlap.R
\name{CORCES_2020.get_ATAC_peak_overlap}
\alias{CORCES_2020.get_ATAC_peak_overlap}
\title{Get overlap between datatable of SNPs and scATAC peaks}
\source{
\url{https://doi.org/10.1038/s41588-020-00721-x}
}
\usage{
CORCES_2020.get_ATAC_peak_overlap(
  finemap_dat,
  FDR_filter = NULL,
  add_cicero = TRUE,
  cell_type_specific = TRUE,
  verbose = TRUE
)
}
\description{
Can optionally add \code{Cicero} coaccessibility scores,
which are also derived from scATAC-seq data.
}
\seealso{
Other CORCES_2020: 
\code{\link{CORCES_2020.get_HiChIP_FitHiChIP_overlap}()},
\code{\link{CORCES_2020.prepare_bulkATAC_peak_overlap}()},
\code{\link{CORCES_2020.prepare_scATAC_peak_overlap}()},
\code{\link{get_CORCES_2020.bulkATACseq_peaks}()},
\code{\link{get_CORCES_2020.cicero_coaccessibility}()},
\code{\link{get_CORCES_2020.hichip_fithichip_loop_calls}()},
\code{\link{get_CORCES_2020.scATACseq_celltype_peaks}()},
\code{\link{get_CORCES_2020.scATACseq_peaks}()}
}
\concept{CORCES_2020}
