% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORCES_2020.get_HiChIP_FitHiChIP_overlap.R
\name{CORCES_2020.get_HiChIP_FitHiChIP_overlap}
\alias{CORCES_2020.get_HiChIP_FitHiChIP_overlap}
\title{Get overlap between data table of SNPs and HiChIP_FitHiChIP
coaccessibility anchors}
\source{
\url{https://doi.org/10.1038/s41588-020-00721-x}
}
\usage{
CORCES_2020.get_HiChIP_FitHiChIP_overlap(finemap_dat, verbose = TRUE)
}
\arguments{
\item{finemap_dat}{Fine-mapping results.}

\item{verbose}{Print messages.}
}
\description{
Anchors are the genomic regions that have evidence of being
functionally connected to one another (coaccessible),
 e.g. enhancer-promoter interactions.
}
\seealso{
Other CORCES_2020: 
\code{\link{CORCES_2020.get_ATAC_peak_overlap}()},
\code{\link{CORCES_2020.prepare_bulkATAC_peak_overlap}()},
\code{\link{CORCES_2020.prepare_scATAC_peak_overlap}()},
\code{\link{get_CORCES_2020.bulkATACseq_peaks}()},
\code{\link{get_CORCES_2020.cicero_coaccessibility}()},
\code{\link{get_CORCES_2020.hichip_fithichip_loop_calls}()},
\code{\link{get_CORCES_2020.scATACseq_celltype_peaks}()},
\code{\link{get_CORCES_2020.scATACseq_peaks}()}
}
\concept{CORCES_2020}
