% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS_bin_plot.R
\name{CS_bin_plot}
\alias{CS_bin_plot}
\title{Plot CS bin counts}
\usage{
CS_bin_plot(merged_DT, show_plot = TRUE)
}
\arguments{
\item{merged_DT}{Merged fine-mapping results data from
\link[echolocatoR]{finemap_loci}.}

\item{show_plot}{Show plot.}
}
\description{
Plot Credible Set bins counts for multiple fine-mapping method results.
}
\examples{
dat <- echodata::BST1 
dat$Locus <- "BST1"
bin_plot <- echoannot::CS_bin_plot(merged_DT = dat)
}
\seealso{
Other summarise: 
\code{\link{CS_counts_plot}()},
\code{\link{peak_overlap_plot}()},
\code{\link{plot_dataset_overlap}()},
\code{\link{super_summary_plot}()}
}
\concept{summarise}
