% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS_counts_plot.R
\name{CS_counts_plot}
\alias{CS_counts_plot}
\title{Bar plot of tool-specific CS sizes}
\usage{
CS_counts_plot(
  merged_DT,
  show_numbers = TRUE,
  ylabel = "Locus",
  legend_nrow = 3,
  label_yaxis = TRUE,
  top_CS_only = FALSE,
  show_plot = TRUE
)
}
\description{
Loci ordered by UCS size (smallest to largest).
}
\examples{
gg_CS <- CS_counts_plot(merged_DT = echodata::get_Nalls2019_merged())
}
\seealso{
Other summarise: 
\code{\link{CS_bin_plot}()},
\code{\link{get_CS_bins}()},
\code{\link{get_CS_counts}()},
\code{\link{get_SNPgroup_counts}()},
\code{\link{peak_overlap_plot}()},
\code{\link{plot_dataset_overlap}()},
\code{\link{results_report}()},
\code{\link{super_summary_plot}()}
}
\concept{summarise}
