% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_chomatin_states.R
\name{filter_chromatin_states}
\alias{filter_chromatin_states}
\title{Filter chromatin states}
\usage{
filter_chromatin_states(
  grl,
  chrom_states = NULL,
  chrom_states_col = "name",
  verbose = TRUE
)
}
\arguments{
\item{grl}{A \link[GenomicRanges]{GRangesList}.}

\item{chrom_states}{Chromatin states to keep. If \code{NULL} (default),
data will not be filtered at all (only annotated).}

\item{chrom_states_col}{Column name in which chromatin state info is stored.}

\item{verbose}{Print messages.}
}
\value{
Annotated/filtered \link[GenomicRanges]{GRangesList}.
}
\description{
Annotate and filter a \link[GenomicRanges]{GRangesList} using
 descriptions of chromatin states from ROADMAP metadata.
}
\keyword{internal}
