% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_consensus_SNPs.R
\name{find_consensus_SNPs}
\alias{find_consensus_SNPs}
\title{Adds fine-mapping summary columns}
\usage{
find_consensus_SNPs(
  finemap_dat,
  credset_thresh = 0.95,
  consensus_thresh = 2,
  sort_by_support = TRUE,
  exclude_methods = NULL,
  top_CS_only = FALSE,
  replace_PP_NAs = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{consensus_thresh}{Threshold for determining
\strong{Consensus_SNP} status.}
}
\description{
Adds several columns that summarise the results across all fine-mapping
tools that were run:
\describe{
\item{Support}{The number of tools in which the SNP was proposed
in a credible set.}
\item{mean.PP}{The mean per-SNP PP across all fine-mapping tools used.}
\item{Consensus_SNP}{Whether or not the SNP was in the credible set of >
 \code{consensus_thresh} SNPs (\emph{default=2}).}
}
}
\examples{
merged_DT <- find_consensus_SNPs(finemap_dat = echodata::get_Nalls2019_merged())
}
\concept{finemapping functions}
\keyword{internal}
