% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_seqinfo.R
\name{fix_seqinfo}
\alias{fix_seqinfo}
\title{Fix bigWig}
\usage{
fix_seqinfo(gr, build = "hg19", verbose = TRUE)
}
\arguments{
\item{gr}{\link[GenomicRanges]{GRanges} object.}

\item{build}{Genome build.}

\item{verbose}{Print messages.}
}
\description{
Fix issues with \link[GenomicRanges]{GRanges} object
imported by \code{rtracklayer::import}, 
Original solution described 
\href{https://www.biostars.org/p/374709/#374727 }{here}.
}
\keyword{internal}
