% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chroms.R
\name{get_chroms}
\alias{get_chroms}
\title{Get chromosomes}
\source{
\href{https://github.com/Bioconductor/BiocGenerics/issues/12}{
GitHub Issues: conflicts with BiocGenerics}
}
\usage{
get_chroms(
  URL,
  chroms = NULL,
  build = "hg38",
  import_format = "bedGraph",
  export_format = import_format,
  save_path = NULL,
  verbose = TRUE
)
}
\description{
Query a bedGraph file by a subset of chromosomes.
This allows you to query whole chromosomes at a time, without needing to
import the entire bedGraph file.
}
\keyword{internal}
