% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CORCES2020_hichip_fithichip_loop_calls.R
\name{get_CORCES2020_hichip_fithichip_loop_calls}
\alias{get_CORCES2020_hichip_fithichip_loop_calls}
\title{FitHiChIP loop calls from human brain tissue}
\source{
\url{https://doi.org/10.1038/s41588-020-00721-x}
}
\usage{
get_CORCES2020_hichip_fithichip_loop_calls()
}
\description{
FitHiChIP loop calls that overlap SNPs derived from analysis
 of H3K27ac HiChIP data.
Each row represents an individual peak identified from the feature
binarization analysis (see methods).
}
\details{
Data originally from \href{https://doi.org/10.1038/s41588-020-00721-x}{
Corces et al. (bioRxiv)}, as of May 2020.
Specifically: \emph{STable10_Coacessibility_Peak_loop_connection},
\emph{HiChIP FitHiChIP Loop Calls} sheet.

\strong{Column dictionary}
\describe{
\item{hg38_Chromosome_Anchor1}{The hg38 chromosome of the first loop Anchor.}
\item{hg38_Start_Anchor1}{The hg38 start position of the first loop Anchor.}
\item{hg38_Stop_Anchor1}{The hg38 stop position of the first loop Anchor.}
\item{Width_Anchor1}{The width of the first loop Anchor.}
\item{hg38_Chromosome_Anchor2}{
The hg38 chromosome of the second loop Anchor.}
\item{hg38_Start_Anchor2}{The hg38 start position of the second loop Anchor.}
\item{hg38_Stop_Anchor2}{The hg38 stop position of the second loop Anchor.}
\item{Width_Anchor2}{The width of the second loop Anchor.}
\item{Score}{The -log10(q-value) of the loop call from FitHiChIP.}
\item{Anchor1_hasSNP}{A boolean variable determining whether the first
anchor overlaps a SNP from our AD/PD GWAS analyses.}
\item{Anchor2_hasSNP}{A boolean variable determining whether the second
anchor overlaps a SNP from our AD/PD GWAS analyses.}
}
}
\examples{
\dontrun{
dat <- readxl::read_excel(
    file.path(
        "~/Desktop/Fine_Mapping/echolocatoR/annotations",
        "Coceres_2020/STable10_Coacessibility_Peak_loop_connection.xlsx"
    ),
    skip = 19, sheet = 1
)
CORCES2020_hichip_fithichip_loop_calls <- data.table::data.table(dat)

#### piggyback ####
tmp <- file.path(tempdir(), "CORCES2020_hichip_fithichip_loop_calls.tsv.gz")
data.table::fwrite(CORCES2020_hichip_fithichip_loop_calls, tmp, sep = "\t")
piggyback::pb_upload(
    file = tmp,
    repo = "RajLabMSSM/echoannot"
)
}
}
\seealso{
Other CORCES2020: 
\code{\link{CORCES2020_get_ATAC_peak_overlap}()},
\code{\link{CORCES2020_get_hichip_fithichip_overlap}()},
\code{\link{CORCES2020_prepare_bulkATAC_peak_overlap}()},
\code{\link{CORCES2020_prepare_scATAC_peak_overlap}()},
\code{\link{get_CORCES2020_bulkATACseq_peaks}()},
\code{\link{get_CORCES2020_cicero_coaccessibility}()},
\code{\link{get_CORCES2020_scATACseq_celltype_peaks}()},
\code{\link{get_CORCES2020_scATACseq_peaks}()}
}
\concept{CORCES2020}
