% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CS_counts.R
\name{get_CS_counts}
\alias{get_CS_counts}
\title{Tally tool-specific and union CS sizes}
\usage{
get_CS_counts(merged_DT, top_CS_only = FALSE)
}
\description{
Tally tool-specific and union CS sizes
}
\examples{
\dontrun{
locus_order <- get_CS_counts(merged_DT = echodata::Nalls2019_merged)
}
}
\seealso{
Other summarise: 
\code{\link{CS_bin_plot}()},
\code{\link{CS_counts_plot}()},
\code{\link{get_CS_bins}()},
\code{\link{get_SNPgroup_counts}()},
\code{\link{peak_overlap_plot}()},
\code{\link{plot_dataset_overlap}()},
\code{\link{results_report}()},
\code{\link{super_summary_plot}()}
}
\concept{summarise}
\keyword{internal}
