% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_histogram_height.R
\name{get_max_histogram_height}
\alias{get_max_histogram_height}
\title{get_max_histogram_height}
\usage{
get_max_histogram_height(gg, round_to = NULL, verbose = TRUE)
}
\description{
get_max_histogram_height
}
\keyword{internal}
