% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NOTT2019_interactome.R
\name{get_NOTT2019_interactome}
\alias{get_NOTT2019_interactome}
\title{Brain cell type-specific enhancers, promoters, and interactomes}
\source{
\url{https://doi.org/10.1126/science.aay0793}

\code{
file <- file.path(
    "~/Desktop/Fine_Mapping/echolocatoR/annotations",
    "NOTT2019/aay0793-Nott-Table-S5.xlsx"
)
sheets <- readxl::excel_sheets(file)
enh_prom_sheets <- grep("enhancers|promoters", sheets, value = TRUE)
other_sheets <- grep("enhancers|promoters", sheets,
    value = TRUE,
    invert = TRUE
)
NOTT2019_interactome <- lapply(other_sheets, function(s) {
    readxl::read_excel(file, sheet = s, skip = 2)
})
NOTT2019_interactome <- append(
    NOTT2019_interactome,
    lapply(enh_prom_sheets, function(s) {
        readxl::read_excel(file,
            sheet = s, skip = 2,
            col_names = c("chr", "start", "end")
        )
    })
)
names(NOTT2019_interactome) <- c(other_sheets, enh_prom_sheets)

#### piggyback ####
tmp <- file.path(tempdir(), "NOTT2019_interactome.rds")
saveRDS(NOTT2019_interactome, tmp)
piggyback::pb_upload(
    file = tmp,
    repo = "RajLabMSSM/echoannot"
)
}
}
\usage{
get_NOTT2019_interactome()
}
\description{
Originally from \href{https://doi.org/10.1126/science.aay0793}{
Nott et al. (2019)}.
Specifically: \emph{aay0793-Nott-Table-S5.xlsx}.
}
\examples{
NOTT2019_interactome <- get_NOTT2019_interactome()
}
\seealso{
Other NOTT2019: 
\code{\link{NOTT2019_bigwig_metadata}},
\code{\link{NOTT2019_epigenomic_histograms}()},
\code{\link{NOTT2019_get_epigenomic_peaks}()},
\code{\link{NOTT2019_get_interactions}()},
\code{\link{NOTT2019_get_interactome}()},
\code{\link{NOTT2019_get_promoter_celltypes}()},
\code{\link{NOTT2019_get_promoter_interactome_data}()},
\code{\link{NOTT2019_get_regulatory_regions}()},
\code{\link{NOTT2019_plac_seq_plot}()},
\code{\link{NOTT2019_superenhancers}()},
\code{\link{get_NOTT2019_superenhancer_interactome}()}
}
\concept{NOTT2019}
