% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NOTT2019_superenhancer_interactome.R
\name{get_NOTT2019_superenhancer_interactome}
\alias{get_NOTT2019_superenhancer_interactome}
\title{Brain cell type-specific interactomes with superenhancers}
\source{
\url{https://science.sciencemag.org/content/366/6469/1134}
}
\usage{
get_NOTT2019_superenhancer_interactome()
}
\description{
Originally from \href{https://science.sciencemag.org/content/366/6469/1134}{
Nott et al. (2019)}.
Specifically: \emph{aay0793-Nott-Table-S6.xlsx}.
}
\examples{
\dontrun{
NOTT2019_superenhancer_interactome <- data.table::data.table(
    readxl::read_excel(
        file.path(
            "~/Desktop/Fine_Mapping/echolocatoR",
            "annotations/NOTT2019/aay0793-Nott-Table-S6.xlsx"
        ),
        skip = 2
    )
)

#### piggyback ####
tmp <- file.path(tempdir(), "NOTT2019_superenhancer_interactome.tsv.gz")
data.table::fwrite(NOTT2019_superenhancer_interactome, tmp, sep = "\t")
piggyback::pb_upload(
    file = tmp,
    repo = "RajLabMSSM/echoannot"
)
}
}
\seealso{
Other NOTT2019: 
\code{\link{NOTT2019_bigwig_metadata}},
\code{\link{NOTT2019_epigenomic_histograms}()},
\code{\link{NOTT2019_get_epigenomic_peaks}()},
\code{\link{NOTT2019_get_interactions}()},
\code{\link{NOTT2019_get_interactome}()},
\code{\link{NOTT2019_get_promoter_celltypes}()},
\code{\link{NOTT2019_get_promoter_interactome_data}()},
\code{\link{NOTT2019_get_regulatory_regions}()},
\code{\link{NOTT2019_plac_seq_plot}()},
\code{\link{NOTT2019_superenhancers}()},
\code{\link{get_NOTT2019_interactome}()}
}
\concept{NOTT2019}
