% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NOTT_2019.superenhancer_interactome.R
\name{get_NOTT_2019.superenhancer_interactome}
\alias{get_NOTT_2019.superenhancer_interactome}
\title{Brain cell type-specific interactomes with superenhancers}
\source{
\url{https://science.sciencemag.org/content/366/6469/1134}
}
\usage{
get_NOTT_2019.superenhancer_interactome()
}
\description{
Originally from \href{https://science.sciencemag.org/content/366/6469/1134}{
Nott et al. (2019)}.
Specifically: \emph{aay0793-Nott-Table-S6.xlsx}.
}
\examples{
\dontrun{
NOTT_2019.superenhancer_interactome <- data.table::data.table(
    readxl::read_excel(
        file.path(
            "~/Desktop/Fine_Mapping/echolocatoR",
            "annotations/Nott_2019/aay0793-Nott-Table-S6.xlsx"
        ),
        skip = 2
    )
)

#### piggyback ####  
tmp <- file.path(tempdir(),"NOTT_2019.superenhancer_interactome.tsv.gz")
data.table::fwrite(NOTT_2019.superenhancer_interactome,tmp,sep="\t")
piggyback::pb_upload(file = tmp,
                     repo = "RajLabMSSM/echoannot")
} 
}
\seealso{
Other NOTT_2019: 
\code{\link{NOTT_2019.bigwig_metadata}},
\code{\link{NOTT_2019.epigenomic_histograms}()},
\code{\link{NOTT_2019.get_epigenomic_peaks}()},
\code{\link{NOTT_2019.get_interactions}()},
\code{\link{NOTT_2019.get_interactome}()},
\code{\link{NOTT_2019.get_promoter_celltypes}()},
\code{\link{NOTT_2019.get_promoter_interactome_data}()},
\code{\link{NOTT_2019.get_regulatory_regions}()},
\code{\link{NOTT_2019.plac_seq_plot}()},
\code{\link{NOTT_2019.superenhancers}()},
\code{\link{get_NOTT_2019.interactome}()}
}
\concept{NOTT_2019}
