% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_consensus_pos.R
\name{get_top_consensus_pos}
\alias{get_top_consensus_pos}
\title{Get top consensus position}
\usage{
get_top_consensus_pos(dat, xvar)
}
\arguments{
\item{dat}{Data}

\item{xvar}{Genomic position column name.}
}
\description{
Get the genomic position of the top fine-mapped Consensus SNP.
}
\keyword{internal}
