% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_window_limits.R
\name{get_window_limits}
\alias{get_window_limits}
\title{Get widow limits}
\usage{
get_window_limits(
  dat,
  index_as_center = TRUE,
  zoom = NULL,
  genomic_units = "Mb",
  verbose = TRUE
)
}
\arguments{
\item{dat}{Data.}

\item{index_as_center}{Use the index/lead SNP 
(the SNP with the smallest P-value) as the center point for the window.}

\item{zoom}{Zoom into the center of the locus when plotting
(without editing the fine-mapping results file).
You can provide either:
\itemize{
\item{The size of your plot window in terms of basepairs
(e.g. \code{zoom=50000} for a 50kb window)}.
\item{How much you want to zoom in (e.g. \code{zoom="1x"}
for the full locus, \code{zoom="2x"}
for 2x zoom into the center of the locus, etc.)}.
}
You can pass a list of window sizes (e.g. \code{c(50000,100000,500000)})
to automatically generate
multiple views of each locus.
This can even be a mix of different style inputs: e.g.
 \code{c("1x","4.5x",25000)}.}

\item{genomic_units}{Which genomic units to return window limits in.}

\item{verbose}{Print messages.}
}
\description{
Get genomic window size limits for a locus plot.
}
\examples{
dat <- echodata::BST1
xlims <- get_window_limits(dat = dat, zoom = 50000)
xlims <- get_window_limits(dat = dat, zoom = "all")
xlims <- get_window_limits(dat = dat, zoom = "5x")
}
\seealso{
Other plot: 
\code{\link{name_filter_convert}()}
}
\concept{plot}
