% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOSHIFTER_utils.R
\name{GOSHIFTER_bed_names}
\alias{GOSHIFTER_bed_names}
\title{Generate ROADMAP BED file names}
\usage{
GOSHIFTER_bed_names(RM_ref, suffix = "_15_coreMarks_mnemonics.bed.gz")
}
\arguments{
\item{RM_ref}{A \code{data.table} of ROADMAP reference entries
(must include an \code{EID} column).}

\item{suffix}{File name suffix
(default \code{"_15_coreMarks_mnemonics.bed.gz"}).}
}
\value{
A character vector of BED file names.
}
\description{
Construct the file names for ROADMAP chromHMM BED files based on
Epigenome IDs in the reference table.
}
\seealso{
Other GOSHIFTER: 
\code{\link{GOSHIFTER}()},
\code{\link{GOSHIFTER_check_overlap}()},
\code{\link{GOSHIFTER_create_LD}()},
\code{\link{GOSHIFTER_create_snpmap}()},
\code{\link{GOSHIFTER_find_folder}()},
\code{\link{GOSHIFTER_get_roadmap_annotations}()},
\code{\link{GOSHIFTER_heatmap}()},
\code{\link{GOSHIFTER_histograms_SNPgroups}()},
\code{\link{GOSHIFTER_histograms_pvals}()},
\code{\link{GOSHIFTER_list_chromatin_states}()},
\code{\link{GOSHIFTER_process_results}()},
\code{\link{GOSHIFTER_run}()},
\code{\link{GOSHIFTER_search_ROADMAP}()},
\code{\link{GOSHIFTER_summarise_results}()}
}
\concept{GOSHIFTER}
\keyword{internal}
