% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOSHIFTER_prepare.R
\name{GOSHIFTER_create_snpmap}
\alias{GOSHIFTER_create_snpmap}
\title{Create GoShifter SNP-map file}
\usage{
GOSHIFTER_create_snpmap(dat, GS_results, verbose = TRUE)
}
\arguments{
\item{dat}{A \code{data.table} or \code{data.frame} with columns
\code{SNP}, \code{CHR}, and \code{POS}.}

\item{GS_results}{Path to the GoShifter results directory where the
\code{snpmap.txt} file will be written.}

\item{verbose}{Print messages (default \code{TRUE}).}
}
\value{
Path to the written \code{snpmap.txt} file (invisibly).
}
\description{
Prepare the tab-delimited SNP-map input file required by GoShifter.
The file contains columns \code{SNP}, \code{Chrom} (prefixed with
\code{"chr"}), and \code{BP}.
}
\examples{
\dontrun{
dat <- echodata::BST1
locus_dir <- echodata::locus_dir
snpmap <- GOSHIFTER_create_snpmap(
    dat = dat,
    GS_results = file.path(locus_dir, "GoShifter")
)
}
}
\seealso{
Other GOSHIFTER: 
\code{\link{GOSHIFTER}()},
\code{\link{GOSHIFTER_bed_names}()},
\code{\link{GOSHIFTER_check_overlap}()},
\code{\link{GOSHIFTER_create_LD}()},
\code{\link{GOSHIFTER_find_folder}()},
\code{\link{GOSHIFTER_get_roadmap_annotations}()},
\code{\link{GOSHIFTER_heatmap}()},
\code{\link{GOSHIFTER_histograms_SNPgroups}()},
\code{\link{GOSHIFTER_histograms_pvals}()},
\code{\link{GOSHIFTER_list_chromatin_states}()},
\code{\link{GOSHIFTER_process_results}()},
\code{\link{GOSHIFTER_run}()},
\code{\link{GOSHIFTER_search_ROADMAP}()},
\code{\link{GOSHIFTER_summarise_results}()}
}
\concept{GOSHIFTER}
\keyword{internal}
