% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOSHIFTER_plot.R
\name{GOSHIFTER_histograms_SNPgroups}
\alias{GOSHIFTER_histograms_SNPgroups}
\title{GoShifter SNP-group histograms}
\usage{
GOSHIFTER_histograms_SNPgroups(GS_groups, show_plot = TRUE)
}
\arguments{
\item{GS_groups}{A \code{data.table} of GoShifter results containing
columns \code{SNP_group}, \code{Annotation}, \code{pval}, and
\code{GS_pval}.}

\item{show_plot}{If \code{TRUE}, print the plot (default \code{TRUE}).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot histograms of GoShifter p-values faceted by SNP group.
}
\examples{
\dontrun{
GS_groups <- data.table::data.table(
    SNP_group = rep(c("A", "B"), each = 5),
    Annotation = "E001",
    pval = runif(10),
    GS_pval = runif(10)
)
GOSHIFTER_histograms_SNPgroups(GS_groups)
}
}
\seealso{
Other GOSHIFTER: 
\code{\link{GOSHIFTER}()},
\code{\link{GOSHIFTER_bed_names}()},
\code{\link{GOSHIFTER_check_overlap}()},
\code{\link{GOSHIFTER_create_LD}()},
\code{\link{GOSHIFTER_create_snpmap}()},
\code{\link{GOSHIFTER_find_folder}()},
\code{\link{GOSHIFTER_get_roadmap_annotations}()},
\code{\link{GOSHIFTER_heatmap}()},
\code{\link{GOSHIFTER_histograms_pvals}()},
\code{\link{GOSHIFTER_list_chromatin_states}()},
\code{\link{GOSHIFTER_process_results}()},
\code{\link{GOSHIFTER_run}()},
\code{\link{GOSHIFTER_search_ROADMAP}()},
\code{\link{GOSHIFTER_summarise_results}()}
}
\concept{GOSHIFTER}
