% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR.name_filter_convert.R
\name{GR.name_filter_convert}
\alias{GR.name_filter_convert}
\title{GR.name_filter_convert}
\usage{
GR.name_filter_convert(GR.final, GR.names, min_hits = 1)
}
\description{
GR.name_filter_convert
}
\seealso{
Other plot: 
\code{\link{PLOT.get_window_limits}()}
}
\concept{plot}
\keyword{internal}
