% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_overlap.R
\name{granges_overlap}
\alias{granges_overlap}
\title{Find overlap between genomic coordinates/ranges}
\usage{
granges_overlap(
  dat1,
  dat2,
  chrom_col.1 = "chrom",
  start_col.1 = "start",
  end_col.1 = "end",
  chrom_col.2 = chrom_col.1,
  start_col.2 = start_col.1,
  end_col.2 = end_col.1,
  return_merged = TRUE,
  chr_format = "NCBI",
  verbose = FALSE
)
}
\description{
Find overlap between genomic coordinates/ranges
}
\concept{GRanges}
\keyword{internal}
