% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_to_bed.R
\name{granges_to_bed}
\alias{granges_to_bed}
\title{Convert GRanges object to BED format and save}
\usage{
granges_to_bed(
  GR.annotations,
  output_path,
  sep = "\\t",
  nThread = 1,
  gzip = FALSE
)
}
\description{
Convert GRanges object to BED format and save
}
\seealso{
Other XGR: 
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.enrichment_bootstrap}()},
\code{\link{XGR.enrichment_plot}()},
\code{\link{XGR.enrichment}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.parse_metadata}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()},
\code{\link{XGR.sep_handler}()}
}
\concept{XGR}
\keyword{internal}
