% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplor_epigenetics_summary.R
\name{haplor_epigenetics_summary}
\alias{haplor_epigenetics_summary}
\title{Summarise \code{HaploR} annotations}
\source{
\href{https://cran.r-project.org/web/packages/haploR/vignettes/haplor-vignette.html}{
HaploR}
}
\usage{
haplor_epigenetics_summary(
  merged_results,
  tissue_list = c("BRN", "BLD"),
  epigenetic_variables = c("Promoter_histone_marks", "Enhancer_histone_marks")
)
}
\description{
Summarise \code{HaploR} annotations
}
\seealso{
Other annotate: 
\code{\link{annotate_missense}()},
\code{\link{biomart_geneInfo}()},
\code{\link{biomart_snp_info}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{haplor_epigenetics_enrichment}()},
\code{\link{haplor_haploreg}()},
\code{\link{haplor_regulomedb}()},
\code{\link{plot_missense}()},
\code{\link{snps_by_mutation_type}()}
}
\concept{annotate}
\keyword{internal}
