% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplor_regulomedb.R
\name{haplor_regulomedb}
\alias{haplor_regulomedb}
\title{Download SNP-wise annotations from RegulomeDB}
\source{
\href{https://cran.r-project.org/web/packages/haploR/vignettes/haplor-vignette.html}{
HaploR}
}
\usage{
haplor_regulomedb(snp_list, verbose = TRUE, chunk_size = NA)
}
\description{
Download SNP-wise annotations from RegulomeDB
}
\seealso{
Other annotate: 
\code{\link{annotate_missense}()},
\code{\link{biomart_geneInfo}()},
\code{\link{biomart_snp_info}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{haplor_epigenetics_enrichment}()},
\code{\link{haplor_epigenetics_summary}()},
\code{\link{haplor_haploreg}()},
\code{\link{plot_missense}()},
\code{\link{snps_by_mutation_type}()}
}
\concept{annotate}
\keyword{internal}
