% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMPACT_files}
\alias{IMPACT_files}
\title{IMPACT files}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 88 rows and 9 columns.
}
\source{
\code{
IMPACT_files <- zenodo_list(token = token)[!endsWith(filename,
                                                     c(".zip",".tbi")),]
usethis::use_data(IMPACT_files, overwrite = TRUE)
}
}
\usage{
data("IMPACT_files")
}
\description{
URLs to annotations/LD-scores generated by
\href{https://github.com/immunogenomics/IMPACT}{IMPACT} 
(Inference and Modeling of Phenotype-related ACtive Transcription),
}
\concept{IMPACT}
\keyword{datasets}
