% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT_process.R
\name{IMPACT_process}
\alias{IMPACT_process}
\title{Process IMPACT files}
\source{
\code{
#### Downloading annotations ####
git clone IMPACT
Mac:
git lfs fetch --all
git lfs pull

Linux:
sudo apt-get install git-lfs
git-lfs fetch --all
git-lfs pull 
}

\code{
#### Installing zen4R ####
## has some system deps that have to be installed beforehand.
sudo apt-get install raptor2-utils
sudo apt-get install rasqal-utils
sudo apt-get install librdf0-dev
}
}
\usage{
IMPACT_process(
  IMPACT_dir,
  sandbox = TRUE,
  title = "IMPACT",
  zipfile = file.path(tempdir(), "IMPACT707"),
  token = Sys.getenv("zenodo_token"),
  validate = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{IMPACT_dir}{Directory where IMPACT repo has been cloned to.}

\item{sandbox}{Whether to use the Zenodo or Zenodo Sandbox API.}

\item{token}{Zenodo
\href{https://zenodo.org/account/settings/applications/tokens/new/}{
Personal access token}.}

\item{verbose}{Print messages.}
}
\description{
Process IMPACT files so that they're all tabix-indexed
and prepare them for upload.
}
\keyword{internal}
