% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT_query.R
\name{IMPACT_query}
\alias{IMPACT_query}
\title{Query IMPACT annotations}
\usage{
IMPACT_query(
  query_dat,
  types = c("annot", "ldscore"),
  populations = c("EAS", "EUR"),
  query_genome = "hg19",
  target_genome = "hg19",
  overlapping_only = TRUE,
  output_format = c("wide", "long", "list"),
  add_metadata = FALSE,
  conda_env = "echoR_mini",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{query_dat}{Variant-level summary statistics.}

\item{types}{File types to include.}

\item{populations}{Population ancestries to include 
("EAS" = East Asian; "EUR" = European).}

\item{query_genome}{Genome build that the \code{query_granges} is aligned to.}

\item{target_genome}{Genome build of the VCF file.}

\item{overlapping_only}{Remove variants that do not overlap with the 
positions in \code{query_dat}.}

\item{output_format}{Output format options:
\itemize{
\item{"wide" : }{Spread annotation across columns and keep 1 row/SNP.}
\item{"long" : }{Melt annotation across rows and allow multiple rows/SNP.}
\item{"list" : }{Do not perform merging of queries and instead return
results as a named list, where the name is 
the file the annotation came from.}
}}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{nThread}{Number of threads to use.}

\item{verbose}{Print messages.}
}
\value{
A named list or data.table of annotations 
merged with \code{query_dat}.
}
\description{
Query annotations/LD-scores generated by
\href{https://github.com/immunogenomics/IMPACT}{IMPACT} 
(Inference and Modeling of Phenotype-related ACtive Transcription),
IMPACT predicts transcription factor (TF) binding at a motif site by 
learning the epigenomic profiles at those sites 
(primarily \href{https://www.encodeproject.org/}{ENCODE}).
All data are aligned to the \emph{hg19} genome build.
All data has also been 
reformatted to tabix indexed files and uploaded to Zenodo 
\href{https://doi.org/10.5281/zenodo.7062238}{here}
 to allow for rapid querying.
}
\examples{
query_dat <- echodata::BST1[1:50,]
annot_dt <- IMPACT_query(query_dat=query_dat, populations="EUR")
}
