% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.bw.filt.R
\name{import.bw.filt}
\alias{import.bw.filt}
\title{Import filtered bigwig}
\usage{
import.bw.filt(bw.file, gr.dat, full_data = TRUE)
}
\arguments{
\item{bw.file}{Path to a bigwig file.}

\item{gr.dat}{GenomicRanges object to query the bigwig file with.}

\item{full_data}{Whether to return the actual read ranges (\code{full_data=T}),
or just the "score" column which summarizes the height of
the aggregated reads across the genome (\code{full_data=TRUE}).}
}
\description{
Import a subset of a bigwig file
based on the coordinates in a GRanges object (gr.dat).
}
\keyword{internal}
