% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_peaks.R
\name{import_peaks}
\alias{import_peaks}
\title{Import peaks}
\usage{
import_peaks(
  ids,
  builds = "hg19",
  query_granges = NULL,
  query_granges_build = NULL,
  query_by_chromosome = FALSE,
  force_new = FALSE,
  cutoff = NULL,
  split_chromosomes = FALSE,
  regex_queries = list(narrowPeak = "narrowpeak", broadPeak = "broadpeak", genericPeak
    = "peak", bedGraph = "bedgraph|graph.gz|bdg.gz", bigWig = "bigwig|bw$"),
  peaks_dir = tempdir(),
  save_path = tempfile(fileext = "grl.hg38.rds"),
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{ids}{Sample IDs from GEO (e.g. "GSM4271282") 
or ENCODE (e.g. "ENCFF048VDO").}

\item{builds}{Genome build that each sample in \code{ids} is aligned to.
This will determine whether whether the \code{query_granges} data need to be 
lifted over to different genome build before querying.
Can be a single character string applied to all \code{ids} (e.g. "hg19"), 
or a vector of the same length as \code{ids} named using the \code{ids} 
(e.g. c("GSM4271282"="hg19", "ENCFF048VDO"="hg38")).}

\item{query_granges}{[Optional]
\link[GenomicRanges]{GRanges} object indicating which genomic
regions to extract from each sample.}

\item{query_granges_build}{[Optional]
Genome build that \code{query_granges} is aligned to.}

\item{force_new}{By default, saved results of the same \code{save_path} name 
will be imported instead of running queries. However you can override this
by setting \code{force_new} to perform new queries regardless and overwrite
 the old \code{save_path} file.}

\item{cutoff}{Cutoff depends on which method you used for score track.
If the file contains pvalue scores from MACS3, score 5 means pvalue 1e-5.
If \code{NULL}, a reasonable \code{cutoff} value will be inferred 
through a \code{cutoff_analysis}.}

\item{split_chromosomes}{Split single-threaded query 
into multi-threaded query across chromosomes. 
This is can be helpful especially when calling peaks from 
large bigWig/bedGraph files.
The number of threads used is set by the \code{nThread} argument.}

\item{regex_queries}{Named list of queries.
Matches will be case-insensitive.}

\item{peaks_dir}{Directory to save peaks to 
(only used when calling peaks from bedGraph files).}

\item{save_path}{Path to save query results to in \emph{.rds} format.}

\item{nThread}{Number of threads to parallelize across.}

\item{verbose}{Print messages.}
}
\value{
A named list of peak files in  \link[GenomicRanges]{GRanges} format.
If peaks could not be recovered for a sample,
 that element will be filled with \code{NA}.
}
\description{
Import pe-computed peak files, or
compute new peaks from bedGraph files.
Automatically detects which database each accession ID is from
 (\href{http://www.ncbi.nlm.nih.gov/geo}{GEO} or 
 \href{https://www.encodeproject.org/}{ENCODE})
and queries a subset of ranges specified in \code{query_granges}.
}
\examples{
grl <- echoannot::import_peaks(ids = "GSM4271282")
}
