% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_peaks_geo.R
\name{import_peaks_geo}
\alias{import_peaks_geo}
\title{Import peaks: GEO}
\usage{
import_peaks_geo(
  gsm,
  build,
  query_granges,
  query_granges_build,
  cutoff = NULL,
  peaks_dir = tempdir(),
  regex_queries = list(narrowPeaks = "narrowpeak", broadPeaks = "broadpeak",
    genericPeaks = "peak", bedGraph = "bedgraph|graph.gz|bdg.gz"),
  verbose = TRUE
)
}
\arguments{
\item{gsm}{GEO GSM id (e.g. "GSM4271282").}

\item{query_granges}{[Optional]
\link[GenomicRanges]{GRanges} object indicating which genomic
regions to extract from each sample.}

\item{query_granges_build}{[Optional]
Genome build that \code{query_granges} is aligned to.}

\item{cutoff}{Cutoff depends on which method you used for score track.
If the file contains pvalue scores from MACS3, score 5 means pvalue 1e-5.
If \code{NULL}, a reasonable \code{cutoff} value will be inferred 
through a \code{cutoff_analysis}.}

\item{peaks_dir}{Directory to save peaks to 
(only used when calling peaks from bedGraph files).}

\item{verbose}{Print messages.}
}
\description{
Import narrow/broad/generic peaks from GEO, or compute peaks with 
\link[echoannot]{call_peaks}.
}
\details{
Must import \link[methods]{new} in my function
because it seems \link[GEOquery]{getGEO} forgot to do this 
(only works when you load the entire \pkg{GEOquery} package first).
}
\keyword{internal}
