% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ucsc_bigwigs.R
\name{import_ucsc_bigwigs}
\alias{import_ucsc_bigwigs}
\title{Import bigwig files from the UCSC Genome Browser}
\usage{
import_ucsc_bigwigs(
  query_dat,
  bigwig_metadata,
  full_data = TRUE,
  xlims = NULL,
  save_path = tempfile(),
  force_new = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{query_dat}{\link[data.table]{data.table} or
 \link[GenomicRanges]{GRanges}
containing genomic coordinates to query the UCSC tracks with.}

\item{bigwig_metadata}{Metadata table with at 
least the following two columns:
\itemize{
\item{"name"}{Unique name of the file.}
\item{"data_link"}{URL to UCSC genome browser bigwig file.}
}}

\item{full_data}{Whether to download the full data 
(genomic ranges of all sequence reads) as opposed
to a reduced representation of the data as a single vector 
(i.e. the aggregated reads "score"). 
Setting \code{full_data=TRUE} is necessary for creating
 histograms and density plots.}

\item{xlims}{Min/max positions to filter from data after initial query.
This helps to capture genomic ranges that only partially overlap with 
\code{query_dat}.}

\item{save_path}{Where to save the processed data.}

\item{force_new}{If a file already exists, download a new one anyway.}

\item{nThread}{Number of threads to parallelise downloads across.}

\item{verbose}{Print messages.}
}
\description{
Import, preprocess, merge, and save data from 
\href{https://genome.ucsc.edu/}{UCSC Genome Browser} tracks.
}
\examples{
bigwig_metadata <- echoannot::NOTT2019_bigwig_metadata[1,]
query_dat = echodata::BST1

bw.gr <- echoannot::import_ucsc_bigwigs(query_dat = query_dat, 
                                        bigwig_metadata = bigwig_metadata)
}
