% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_finemapping_results.R
\name{merge_finemapping_results}
\alias{merge_finemapping_results}
\title{Merge fine-mapping results from all loci}
\usage{
merge_finemapping_results(
  dataset = file.path(tempdir(), "Data/GWAS"),
  minimum_support = 1,
  include_leadSNPs = TRUE,
  LD_reference = NULL,
  save_path = tempfile(fileext = "merged_results.csv.gz"),
  from_storage = TRUE,
  haploreg_annotation = FALSE,
  regulomeDB_annotation = FALSE,
  biomart_annotation = FALSE,
  credset_thresh = 0.95,
  consensus_thresh = 2,
  exclude_methods = NULL,
  top_CS_only = FALSE,
  verbose = TRUE,
  nThread = 1
)
}
\arguments{
\item{dataset}{Path to the folder you want to recursively search
for results files within
 (e.g. \url{"Data/GWAS/Nalls23andMe_2019"}).
Set this to a path that includes multiple subfolders if you want
to gather results
from multiple studies at once
(e.g. \url{"Data/GWAS"}).}

\item{minimum_support}{Filter SNPs by the minimum number
of fine-mapping tools that contained the SNP in their Credible Set.}

\item{include_leadSNPs}{Include lead GWAS/QTL SNPs per locus
(regardless of other filtering criterion).}

\item{LD_reference}{LD reference to use:
\itemize{
\item{"1KGphase1" : }{1000 Genomes Project Phase 1}
\item{"1KGphase3" : }{1000 Genomes Project Phase 3}
\item{"UKB" : }{Pre-computed LD from a British
European-decent subset of UK Biobank.}
\item{"<vcf_path>" : }{User-supplied path to a custom VCF file 
to compute LD matrix from.}
}}

\item{save_path}{Path to save merged table to.}

\item{from_storage}{Search for stored results files.}

\item{haploreg_annotation}{Annotate SNPs with HaploReg
(using \code{HaploR}).}

\item{regulomeDB_annotation}{Annotate SNPs with regulomeDB
(using \code{HaploR}).}

\item{biomart_annotation}{Annotate SNPs with \code{biomart}.}

\item{credset_thresh}{The minimum mean Posterior Probability
(across all fine-mapping methods used) of SNPs to be included in 
the "mean.CS" column.}

\item{consensus_thresh}{The minimum number of fine-mapping tools 
in which a SNP is in the Credible Set in order to be included in the 
"Consensus_SNP" column.}

\item{exclude_methods}{Exclude certain fine-mapping methods when estimating
\strong{mean.CS} and \strong{Consensus_SNP}.}

\item{top_CS_only}{Only include the top 1 CS per fine-mapping method.}

\item{verbose}{Print messages.}

\item{nThread}{Number of threads to parallelise across.}
}
\description{
Gather fine-mapping results from \pkg{echolocatoR} across all loci
and merge into a single data.frame.
}
\examples{
dataset <- echodata::get_Nalls2019_loci(return_dir = TRUE)
merged_DT <- echoannot:: merge_finemapping_results(dataset = dataset)
}
