% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOTIFBREAKR_filter.R
\name{MOTIFBREAKR_filter}
\alias{MOTIFBREAKR_filter}
\title{Merge and filter \pkg{motifbreakR} + \pkg{echolocatoR} results}
\usage{
MOTIFBREAKR_filter(
  mb_res,
  merged_DT,
  filter_by_locus = NULL,
  remove_NA_geneSymbol = TRUE,
  pct_threshold = NULL,
  pvalue_threshold = 1e-04,
  qvalue_threshold = 0.05,
  effect_strengths = NULL,
  snp_filter = "Support>0",
  top_geneSymbol_hits = NULL,
  no_no_loci = NULL,
  verbose = TRUE
)
}
\arguments{
\item{mb_res}{Results generated by \link[echoannot]{MOTIFBREAKR}, in 
\link[GenomicRanges]{GRanges} format.}

\item{merged_DT}{Table with columns \emph{Locus} and 
\emph{SNP} to filter \code{mb_res} by.}

\item{filter_by_locus}{Filter \code{mb_res} 
to only include SNPs present in a given \emph{Locus} (e.g. \code{"BST1"}).
Set to \code{NULL} (default) for not perform any filtering.
Requires \code{merged_DT} argument to be supplied.}

\item{remove_NA_geneSymbol}{Remove results where \code{geneSymbol==NA}.}

\item{pct_threshold}{Remove rows below the percentage of 
the optimal binding score (PCT) threshold.}

\item{pvalue_threshold}{Remove rows below the raw
significance value (p-value) threshold.}

\item{qvalue_threshold}{Remove rows below the multiple testing-corrected 
significance value (q-value) threshold.}

\item{effect_strengths}{Only include results with certain effect strengths.}

\item{snp_filter}{Condition to filter SNPs by, 
after \code{mb_res} and \code{merged_DT} have been merged together 
into one table.}

\item{top_geneSymbol_hits}{Only include N top results per gene symbol 
based on absolute \code{risk_score}, where N=\code{top_geneSymbol_hits}.
If \code{top_geneSymbol_hits=NULL}, no such filtering is performed.}

\item{no_no_loci}{Filter out SNPs contained within specific loci
in the \code{merged_DT} table.}

\item{verbose}{Print messages.}
}
\value{
\link[data.table]{data.table} 
with the filtered motif disruption results 
after merging with \code{merged_DT}.
}
\description{
For each SNP we have at least one allele achieving a p-value<1e-4 threshold
 that we required.
The \emph{seqMatch} column shows what the reference genome sequence 
is at that location,
with the variant position appearing in an uppercase letter.
\emph{pctRef} and \emph{pctAlt} display the the score for the motif 
in the sequence as a percentage of the best score that motif 
could achieve on an ideal sequence.
In other words, \emph{(scoreVariant−minscorePWM)/(maxscorePWM−minscorePWM)}.
We can also see the absolute scores for our method in
 \emph{scoreRef} and \emph{scoreAlt} and their respective p-values.
}
\examples{
merged_DT <- echodata::get_Nalls2019_merged()
mb_res <- MOTIFBREAKR(rsid_list = c("rs11175620"),
                      # limit the number of datasets tests 
                      # for demonstration purposes only
                      pwmList_max = 4,
                      calculate_pvals = FALSE)
mb_res_filt <- MOTIFBREAKR_filter(mb_res = mb_res,
                                  merged_DT = merged_DT)
}
