% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOTIFBREAKR_filter_by_metadata.R
\name{MOTIFBREAKR_filter_by_metadata}
\alias{MOTIFBREAKR_filter_by_metadata}
\title{Filter \pkg{motifbreakR} results}
\usage{
MOTIFBREAKR_filter_by_metadata(mb_res, Organism = "Hsapiens")
}
\arguments{
\item{mb_res}{Results generated by \link[echoannot]{MOTIFBREAKR}, in 
\link[GenomicRanges]{GRanges} format.}
}
\description{
Filter by motif database metadata.
}
\seealso{
Other motifbreakR: 
\code{\link{MOTIFBREAKR_summarize}()},
\code{\link{MOTIFBREAKR}()}
}
\concept{motifbreakR}
\keyword{internal}
