% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOTIFBREAKR_make_id.R
\name{MOTIFBREAKR_make_id}
\alias{MOTIFBREAKR_make_id}
\title{MOTIFBREAKR: make ID}
\usage{
MOTIFBREAKR_make_id(mb_res)
}
\arguments{
\item{mb_res}{Results generated by \link[echoannot]{MOTIFBREAKR}, in 
\link[GenomicRanges]{GRanges} format.}
}
\value{
Motif disruption predictions in 
 \link[GenomicRanges]{GRanges} format.
}
\description{
MOTIFBREAKR support function.
}
\keyword{internal}
