% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOTIFBREAKR_plot.R
\name{MOTIFBREAKR_plot}
\alias{MOTIFBREAKR_plot}
\title{Plot \pkg{motifbreakR} results}
\source{
\href{https://pubmed.ncbi.nlm.nih.gov/26272984/}{Publication}
\href{https://github.com/Simon-Coetzee/MotifBreakR}{GitHub}
\href{https://github.com/Simon-Coetzee/motifBreakR/issues/31}{Viewport bug}
}
\usage{
MOTIFBREAKR_plot(
  mb_res,
  mb_filter = NULL,
  rsid = NULL,
  effect = c("strong", "weak"),
  results_dir = file.path(tempdir(), "results"),
  height = 3,
  width = 7,
  verbose = TRUE
)
}
\arguments{
\item{mb_res}{Results generated by \link[echoannot]{MOTIFBREAKR}, in 
\link[GenomicRanges]{GRanges} format.}

\item{mb_filter}{\link[GenomicRanges]{GRanges} object 
used to filter \code{mb_res} before plotting.}

\item{rsid}{Character; the identifier of the variant to be visualized}

\item{effect}{Character; show motifs that are strongly effected \code{c("strong")},
weakly effected \code{c("weak")}, or both \code{c("strong", "weak")}}

\item{results_dir}{Directory where results should be saved 
as a file named:
 \emph{<results_dir>/_genome_wide/motifbreakR/motifbreakR_results.rds}.
If \code{NULL}, results will not be saved to disk.}

\item{height}{the height of the device.}

\item{width}{the width of the device.}

\item{verbose}{Print messages.}
}
\value{
Named list of motif plot paths.
}
\description{
Plot motif disruption results generated by \link[echoannot]{MOTIFBREAKR}.
}
\details{
\emph{Notes:}\cr
\itemize{
\item{Saving as a PDF seems to work much better than PNG format 
(at least when using \pkg{grDevices}).
}
}
}
\examples{
library(BSgenome) ## <-- IMPORTANT!
library(BSgenome.Hsapiens.UCSC.hg19) ## <-- IMPORTANT!
#### Example fine-mapping results ####
merged_DT <- echodata::get_Nalls2019_merged()
#### Run motif analyses ####
mb_res <- MOTIFBREAKR(rsid_list = c("rs11175620"),
                      # limit the number of datasets tested 
                      # for demonstration purposes only
                      pwmList_max = 5,
                      calculate_pvals = FALSE)  
plot_paths <- MOTIFBREAKR_plot(mb_res = mb_res) 
}
