% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOTIFBREAKR_summarize.R
\name{MOTIFBREAKR_summarize}
\alias{MOTIFBREAKR_summarize}
\title{Summarize \pkg{motifbreakR} + \pkg{echolocatoR} results}
\usage{
MOTIFBREAKR_summarize(mb_merge)
}
\arguments{
\item{mb_merge}{Output of \link[echoannot]{MOTIFBREAKR_filter}.}
}
\value{
Per-locus summary in data.frame format.
}
\description{
Summarize \pkg{motifbreakR} + \pkg{echolocatoR} results after they have 
been merged together with \link[echoannot]{MOTIFBREAKR_filter}.
}
\examples{
merged_DT <- echodata::get_Nalls2019_merged()
mb_res <- MOTIFBREAKR(rsid_list = c("rs11175620"),
                      # limit the number of datasets tests 
                      # for demonstration purposes only
                      pwmList_max = 4,
                      calculate_pvals = FALSE)
mb_merge <- MOTIFBREAKR_filter(mb_res = mb_res,
                               merged_DT = merged_DT,
                               pvalue_threshold = NULL)
summary_ls <- MOTIFBREAKR_summarize(mb_merge = mb_merge)                         
}
\seealso{
Other motifbreakR: 
\code{\link{MOTIFBREAKR_filter_by_metadata}()},
\code{\link{MOTIFBREAKR}()}
}
\concept{motifbreakR}
