% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_filter_convert.R
\name{name_filter_convert}
\alias{name_filter_convert}
\title{name_filter_convert}
\usage{
name_filter_convert(grl, min_hits = 1)
}
\description{
name_filter_convert
}
\seealso{
Other plot: 
\code{\link{get_window_limits}()}
}
\concept{plot}
\keyword{internal}
