% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOTT2019_bigwig_metadata}
\alias{NOTT2019_bigwig_metadata}
\title{Metadata and links to data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 18 rows and 14 columns.
}
\source{
\url{https://doi.org/10.1126/science.aay0793}

\code{
NOTT2019_bigwig_metadata <- data.table::data.table(
    readxl::read_excel(
        file.path(
            "~/Desktop/Fine_Mapping/echolocatoR/annotations",
            "NOTT2019/NOTT2019_snEpigenomics.xlsx"
        )
    )
)
usethis::use_data(NOTT2019_bigwig_metadata, overwrite = TRUE)
}
}
\usage{
data("NOTT2019_bigwig_metadata")
}
\description{
Metadata for cell type-specific epigenomic bigWig files hosted
 on UCSC Genome Browser.
bigWig files contain the genomic ranges from each epigenomic assay,
as well as a Score column which describes the peaks of the aggregate reads.
}
\seealso{
Other NOTT2019: 
\code{\link{NOTT2019_epigenomic_histograms}()},
\code{\link{NOTT2019_get_epigenomic_peaks}()},
\code{\link{NOTT2019_get_interactions}()},
\code{\link{NOTT2019_get_interactome}()},
\code{\link{NOTT2019_get_promoter_celltypes}()},
\code{\link{NOTT2019_get_promoter_interactome_data}()},
\code{\link{NOTT2019_get_regulatory_regions}()},
\code{\link{NOTT2019_plac_seq_plot}()},
\code{\link{NOTT2019_superenhancers}()},
\code{\link{get_NOTT2019_interactome}()},
\code{\link{get_NOTT2019_superenhancer_interactome}()}
}
\concept{NOTT2019}
\keyword{datasets}
