% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOTT2019_epigenomic_histograms.R
\name{NOTT2019_epigenomic_histograms}
\alias{NOTT2019_epigenomic_histograms}
\title{Plot brain cell-specific epigenomic data}
\source{
\href{https://doi.org/10.1126/science.aay0793}{Nott et al. (2019)}
}
\usage{
NOTT2019_epigenomic_histograms(
  dat,
  bigwig_metadata = echoannot::NOTT2019_bigwig_metadata,
  locus_dir = tempdir(),
  show_plot = TRUE,
  save_plot = FALSE,
  full_data = TRUE,
  return_assay_track = FALSE,
  binwidth = 200,
  density_adjust = 0.2,
  zoom = "1x",
  strip.text.y.angle = 90,
  xtext = TRUE,
  geom = "density",
  plot_formula = "Cell_type ~.",
  fill_var = "Assay",
  genomic_units = "Mb",
  as_ggplot = TRUE,
  dpi = 300,
  height = 15,
  width = 8,
  nThread = 1,
  save_annot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Fine-mapping results data from \link[echolocatoR]{finemap_loci}.}

\item{bigwig_metadata}{Metadata table with at 
least the following two columns:
\itemize{
\item{"name"}{Unique name of the file.}
\item{"data_link"}{URL to UCSC genome browser bigwig file.}
}}

\item{locus_dir}{Locus-specific directory.}

\item{show_plot}{Show plot.}

\item{save_plot}{Whether to save the plot.}

\item{full_data}{Whether to download the full data 
(genomic ranges of all sequence reads) as opposed
to a reduced representation of the data as a single vector 
(i.e. the aggregated reads "score"). 
Setting \code{full_data=TRUE} is necessary for creating
 histograms and density plots.}

\item{return_assay_track}{Return only the assay track 
(before adding the rest of the tracks and showing the plot).}

\item{binwidth}{
    width of the bins.
  }

\item{density_adjust}{Passed to \code{adjust} argument in 
\link[ggplot2]{geom_density}.}

\item{zoom}{Zoom into the center of the locus when plotting
(without editing the fine-mapping results file).
You can provide either:
\itemize{
\item{The size of your plot window in terms of basepairs
(e.g. \code{zoom=50000} for a 50kb window)}.
\item{How much you want to zoom in (e.g. \code{zoom="1x"}
for the full locus, \code{zoom="2x"}
for 2x zoom into the center of the locus, etc.)}.
}
You can pass a list of window sizes (e.g. \code{c(50000,100000,500000)})
to automatically generate
multiple views of each locus.
This can even be a mix of different style inputs: e.g.
 \code{c("1x","4.5x",25000)}.}

\item{strip.text.y.angle}{Angle of the y-axis facet labels.}

\item{xtext}{Whether to include x-axis title and text.}

\item{geom}{
    Geom to use (Single character for now). Please see section Geometry
    for details.
  }

\item{plot_formula}{Formula passed to \code{facets} argument in 
\link[ggplot2]{facet_grid}.}

\item{fill_var}{Variable name to use for plot \code{fill} argument.}

\item{genomic_units}{Which genomic units to return window limits in.}

\item{as_ggplot}{Return plot as \code{ggplot2} 
(\code{TRUE}) or \code{Tracks} (\code{FALSE}) object.}

\item{dpi}{dpi to use for raster graphics}

\item{height}{height (defaults to the height of current
  plotting window)}

\item{width}{width (defaults to the width of current
  plotting window)}

\item{nThread}{Number of threads to parallelise downloads across.}

\item{save_annot}{Save the queried subset of bigwig annotations.}

\item{verbose}{Print messages.}
}
\description{
Brain cell-specific epigenomic data from Nott et al. (2019).
}
\examples{
nott2019_track <- echoannot::NOTT2019_epigenomic_histograms(
    dat = echodata::BST1, 
    bigwig_metadata = echoannot::NOTT2019_bigwig_metadata[1:2,])
}
\seealso{
Other NOTT2019: 
\code{\link{NOTT2019_bigwig_metadata}},
\code{\link{NOTT2019_get_epigenomic_peaks}()},
\code{\link{NOTT2019_get_interactions}()},
\code{\link{NOTT2019_get_interactome}()},
\code{\link{NOTT2019_get_promoter_celltypes}()},
\code{\link{NOTT2019_get_promoter_interactome_data}()},
\code{\link{NOTT2019_get_regulatory_regions}()},
\code{\link{NOTT2019_plac_seq_plot}()},
\code{\link{NOTT2019_superenhancers}()},
\code{\link{get_NOTT2019_interactome}()},
\code{\link{get_NOTT2019_superenhancer_interactome}()}
}
\concept{NOTT2019}
