% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOTT2019_get_epigenomic_peaks.R
\name{NOTT2019_get_epigenomic_peaks}
\alias{NOTT2019_get_epigenomic_peaks}
\title{Download cell type-specific epigenomic peaks}
\source{
\href{https://science.sciencemag.org/content/366/6469/1134}{Nott et al. (2019)}
\url{https://genome.ucsc.edu/cgi-bin/hgTracks?db=hg19&lastVirtModeType=default&lastVirtModeExtraState=&virtModeType=default&virtMode=0&nonVirtPosition=&position=chr2:127770344-127983251&hgsid=778249165_ySowqECRKNxURRn6bafH0yewAiuf}
}
\usage{
NOTT2019_get_epigenomic_peaks(
  assays = c("ATAC", "H3K27ac", "H3K4me3"),
  cell_types = c("neurons", "microglia", "oligo", "astrocytes"),
  convert_to_granges = TRUE,
  save_dir = tempdir(),
  force_new = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{assays}{Which epigenomic assays to import data from.}

\item{cell_types}{Which cell-types to import data from.}

\item{convert_to_granges}{Whether to convert the peaks to a 
\link[GenomicRanges]{GRanges} object.}

\item{save_dir}{Where to save the processed data.}

\item{force_new}{If the saved data already exists, re-downloaded anyway.}

\item{nThread}{Number of threads to parallelise downloads across.}

\item{verbose}{Print messages.}
}
\description{
API access to brain cell type-specific epigenomic peaks (bed format)
from Nott et al. (2019).
}
\examples{
PEAKS <- echoannot::NOTT2019_get_epigenomic_peaks() 
}
\seealso{
Other NOTT2019: 
\code{\link{NOTT2019_bigwig_metadata}},
\code{\link{NOTT2019_epigenomic_histograms}()},
\code{\link{NOTT2019_get_interactions}()},
\code{\link{NOTT2019_get_interactome}()},
\code{\link{NOTT2019_get_promoter_celltypes}()},
\code{\link{NOTT2019_get_promoter_interactome_data}()},
\code{\link{NOTT2019_get_regulatory_regions}()},
\code{\link{NOTT2019_plac_seq_plot}()},
\code{\link{NOTT2019_superenhancers}()},
\code{\link{get_NOTT2019_interactome}()},
\code{\link{get_NOTT2019_superenhancer_interactome}()}
}
\concept{NOTT2019}
