% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOTT2019_get_interactome.R
\name{NOTT2019_get_interactome}
\alias{NOTT2019_get_interactome}
\title{Import cell type-specific interactomes}
\source{
\href{https://doi.org/10.1126/science.aay0793}{Nott et al. (2019)}
}
\usage{
NOTT2019_get_interactome(
  annot_sub,
  top.consensus.pos,
  marker_key = NOTT2019_marker_key(),
  verbose = TRUE
)
}
\arguments{
\item{annot_sub}{annot_sub}

\item{top.consensus.pos}{Top Consensus SNP position.}

\item{marker_key}{Marker key.}

\item{verbose}{Print messages.}
}
\description{
Brain cell-specific epigenomic data from Nott et al. (2019).
}
\examples{
dat <- echodata::BST1
annot_sub <- NOTT2019_get_promoter_interactome_data(dat = dat)
interact.DT <- NOTT2019_get_interactome(annot_sub=annot_sub, 
                                        top.consensus.pos = 15712787)
}
\seealso{
Other NOTT2019: 
\code{\link{NOTT2019_bigwig_metadata}},
\code{\link{NOTT2019_epigenomic_histograms}()},
\code{\link{NOTT2019_get_epigenomic_peaks}()},
\code{\link{NOTT2019_get_interactions}()},
\code{\link{NOTT2019_get_promoter_celltypes}()},
\code{\link{NOTT2019_get_promoter_interactome_data}()},
\code{\link{NOTT2019_get_regulatory_regions}()},
\code{\link{NOTT2019_plac_seq_plot}()},
\code{\link{NOTT2019_superenhancers}()},
\code{\link{get_NOTT2019_interactome}()},
\code{\link{get_NOTT2019_superenhancer_interactome}()}
}
\concept{NOTT2019}
