% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOTT2019_get_promoter_celltypes.R
\name{NOTT2019_get_promoter_celltypes}
\alias{NOTT2019_get_promoter_celltypes}
\title{Get promoter cell types}
\usage{
NOTT2019_get_promoter_celltypes(annot_sub, marker_key, verbose = TRUE)
}
\description{
Brain cell-specific epigenomic data from Nott et al. (2019).
}
\seealso{
Other NOTT2019: 
\code{\link{NOTT2019_bigwig_metadata}},
\code{\link{NOTT2019_epigenomic_histograms}()},
\code{\link{NOTT2019_get_epigenomic_peaks}()},
\code{\link{NOTT2019_get_interactions}()},
\code{\link{NOTT2019_get_interactome}()},
\code{\link{NOTT2019_get_promoter_interactome_data}()},
\code{\link{NOTT2019_get_regulatory_regions}()},
\code{\link{NOTT2019_plac_seq_plot}()},
\code{\link{NOTT2019_superenhancers}()},
\code{\link{get_NOTT2019_interactome}()},
\code{\link{get_NOTT2019_superenhancer_interactome}()}
}
\concept{NOTT2019}
\keyword{internal}
