% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOTT2019_get_regulatory_regions.R
\name{NOTT2019_get_regulatory_regions}
\alias{NOTT2019_get_regulatory_regions}
\title{Get regulatory regions: Nott2019}
\source{
\href{https://doi.org/10.1126/science.aay0793}{Nott et al. (2019)}
}
\usage{
NOTT2019_get_regulatory_regions(
  as_granges = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{as_granges}{Return results a \link[GenomicRanges]{GRanges} object
instead of a \link[data.table]{data.table}.}

\item{nThread}{Number of threads to parallelize across.}

\item{verbose}{Print messages.}
}
\description{
Get epigenomic peak ranges from Nott2019.
}
\examples{
regions <- NOTT2019_get_regulatory_regions()
}
\seealso{
Other NOTT2019: 
\code{\link{NOTT2019_bigwig_metadata}},
\code{\link{NOTT2019_epigenomic_histograms}()},
\code{\link{NOTT2019_get_epigenomic_peaks}()},
\code{\link{NOTT2019_get_interactions}()},
\code{\link{NOTT2019_get_interactome}()},
\code{\link{NOTT2019_get_promoter_celltypes}()},
\code{\link{NOTT2019_get_promoter_interactome_data}()},
\code{\link{NOTT2019_plac_seq_plot}()},
\code{\link{NOTT2019_superenhancers}()},
\code{\link{get_NOTT2019_interactome}()},
\code{\link{get_NOTT2019_superenhancer_interactome}()}
}
\concept{NOTT2019}
