% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOTT2019_plac_seq_plot.R
\name{NOTT2019_plac_seq_plot}
\alias{NOTT2019_plac_seq_plot}
\title{Plot brain cell-specific interactome data}
\source{
\href{https://science.sciencemag.org/content/366/6469/1134}{Nott et al. (2019)}
\url{https://genome.ucsc.edu/cgi-bin/hgTracks?db=hg19&lastVirtModeType=default&lastVirtModeExtraState=&virtModeType=default&virtMode=0&nonVirtPosition=&position=chr2:127770344-127983251&hgsid=778249165_ySowqECRKNxURRn6bafH0yewAiuf}
}
\usage{
NOTT2019_plac_seq_plot(
  dat = NULL,
  locus_dir = NULL,
  title = NULL,
  print_plot = TRUE,
  save_plot = TRUE,
  return_interaction_track = FALSE,
  x_limits = NULL,
  zoom_window = NULL,
  index_SNP = NULL,
  genomic_units = "POS",
  color_dict = c(enhancers = "springgreen2", promoters = "purple", anchors = "black"),
  return_consensus_overlap = TRUE,
  show_arches = TRUE,
  highlight_plac = TRUE,
  show_regulatory_rects = TRUE,
  show_anchors = TRUE,
  strip.text.y.angle = 0,
  xtext = TRUE,
  save_annot = FALSE,
  point_size = 2,
  height = 7,
  width = 7,
  dpi = 300,
  return_as = "Tracks",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{x_limits}{x-axis limits to be applied to all plots
(useful when trying to keep a common coordinate system).}

\item{highlight_plac}{Whether to scale opacity of PLAC-seq interactions 
(arches) such that interactions with anchors containing Consensus SNPs 
will be colored darker (Default: \code{TRUE}).
If \code{FALSE}, will instead apply the same opacity level
 to all interactions.
 @param genomic_units The genomic units that should be used for the x-axis.}

\item{return_as}{Plot class to convert \code{plot_list} to:
\itemize{
\item{"ggplot"}{\link[ggplot2]{ggplot}}
\item{"ggbio"}{\link[ggbio]{ggbio}}
\item{"patchwork"}{\pkg{patchwork}}
\item{"Tracks"}{\link[ggbio]{tracks}}
\item{NULL}{Return original object.}
}}

\item{verbose}{Print messages.}
}
\description{
Plot brain cell-specific interactome data from
 \href{https://science.sciencemag.org/content/366/6469/1134}{Nott et al. (2019)}.
}
\examples{
trks_plus_lines <- echoannot::NOTT2019_plac_seq_plot(dat = echodata::BST1) 
}
\seealso{
Other NOTT2019: 
\code{\link{NOTT2019_bigwig_metadata}},
\code{\link{NOTT2019_epigenomic_histograms}()},
\code{\link{NOTT2019_get_epigenomic_peaks}()},
\code{\link{NOTT2019_get_interactions}()},
\code{\link{NOTT2019_get_interactome}()},
\code{\link{NOTT2019_get_promoter_celltypes}()},
\code{\link{NOTT2019_get_promoter_interactome_data}()},
\code{\link{NOTT2019_get_regulatory_regions}()},
\code{\link{NOTT2019_superenhancers}()},
\code{\link{get_NOTT2019_interactome}()},
\code{\link{get_NOTT2019_superenhancer_interactome}()}
}
\concept{NOTT2019}
