% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOTT_2019.epigenomic_histograms.R
\name{NOTT_2019.epigenomic_histograms}
\alias{NOTT_2019.epigenomic_histograms}
\title{Plot brain cell-specific epigenomic data}
\source{
\href{https://science.sciencemag.org/content/366/6469/1134}{Nott et al. (2019)}
\url{https://genome.ucsc.edu/cgi-bin/hgTracks?db=hg19&lastVirtModeType=default&lastVirtModeExtraState=&virtModeType=default&virtMode=0&nonVirtPosition=&position=chr2:127770344-127983251&hgsid=778249165_ySowqECRKNxURRn6bafH0yewAiuf}
}
\usage{
NOTT_2019.epigenomic_histograms(
  finemap_dat,
  locus_dir,
  show_plot = TRUE,
  save_plot = TRUE,
  full_data = TRUE,
  return_assay_track = FALSE,
  binwidth = 200,
  density_adjust = 0.2,
  plot.zoom = "1x",
  strip.text.y.angle = 90,
  xtext = TRUE,
  geom = "density",
  plot_formula = "Cell_type ~.",
  fill_var = "Assay",
  bigwig_dir = NULL,
  genomic_units = "Mb",
  as_ggplot = TRUE,
  nThread = 1,
  save_annot = FALSE,
  verbose = TRUE
)
}
\description{
Brain cell-specific epigenomic data from Nott et al. (2019).
}
\examples{
track.Nott_histo <- NOTT_2019.epigenomic_histograms(
    finemap_dat = echodata::BST1,
    locus_dir = echodata::locus_dir,
    save_plot = FALSE,
    return_assay_track = TRUE,
    save_annot = FALSE
)
}
\seealso{
Other NOTT_2019: 
\code{\link{NOTT_2019.bigwig_metadata}},
\code{\link{NOTT_2019.get_epigenomic_peaks}()},
\code{\link{NOTT_2019.get_interactions}()},
\code{\link{NOTT_2019.get_interactome}()},
\code{\link{NOTT_2019.get_promoter_celltypes}()},
\code{\link{NOTT_2019.get_promoter_interactome_data}()},
\code{\link{NOTT_2019.get_regulatory_regions}()},
\code{\link{NOTT_2019.plac_seq_plot}()},
\code{\link{NOTT_2019.superenhancers}()},
\code{\link{get_NOTT_2019.interactome}()},
\code{\link{get_NOTT_2019.superenhancer_interactome}()}
}
\concept{NOTT_2019}
\keyword{internal}
