% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOTT_2019.plac_seq_plot.R
\name{NOTT_2019.plac_seq_plot}
\alias{NOTT_2019.plac_seq_plot}
\title{Plot brain cell-specific interactome data}
\source{
\href{https://science.sciencemag.org/content/366/6469/1134}{Nott et al. (2019)}
\url{https://genome.ucsc.edu/cgi-bin/hgTracks?db=hg19&lastVirtModeType=default&lastVirtModeExtraState=&virtModeType=default&virtMode=0&nonVirtPosition=&position=chr2:127770344-127983251&hgsid=778249165_ySowqECRKNxURRn6bafH0yewAiuf}
}
\usage{
NOTT_2019.plac_seq_plot(
  finemap_dat = NULL,
  locus_dir = NULL,
  title = NULL,
  print_plot = TRUE,
  save_plot = TRUE,
  return_interaction_track = FALSE,
  xlims = NULL,
  zoom_window = NULL,
  index_SNP = NULL,
  genomic_units = "Mb",
  color_dict = c(enhancers = "springgreen2", promoters = "purple", anchors = "black"),
  return_consensus_overlap = TRUE,
  show_arches = TRUE,
  highlight_plac = FALSE,
  show_regulatory_rects = TRUE,
  show_anchors = TRUE,
  strip.text.y.angle = 0,
  xtext = TRUE,
  save_annot = FALSE,
  point_size = 2,
  height = 7,
  width = 7,
  dpi = 300,
  as_ggplot = TRUE,
  nThread = 1,
  verbose = TRUE
)
}
\description{
Plot brain cell-specific interactome data
}
\examples{
\dontrun{
data("BST1")
data("locus_dir")
trks_plus_lines <- NOTT_2019.plac_seq_plot(finemap_dat = BST1, locus_dir = file.path("~/Desktop", locus_dir), highlight_plac = TRUE)
# Zoom in
trks_plus_lines <- NOTT_2019.plac_seq_plot(finemap_dat = BST1, locus_dir = file.path("~/Desktop", locus_dir), zoom_window = 500000, highlight_plac = TRUE)
}
}
\seealso{
Other NOTT_2019: 
\code{\link{NOTT_2019.bigwig_metadata}},
\code{\link{NOTT_2019.epigenomic_histograms}()},
\code{\link{NOTT_2019.get_epigenomic_peaks}()},
\code{\link{NOTT_2019.get_interactions}()},
\code{\link{NOTT_2019.get_interactome}()},
\code{\link{NOTT_2019.get_promoter_celltypes}()},
\code{\link{NOTT_2019.get_promoter_interactome_data}()},
\code{\link{NOTT_2019.get_regulatory_regions}()},
\code{\link{NOTT_2019.superenhancers}()},
\code{\link{get_NOTT_2019.interactome}()},
\code{\link{get_NOTT_2019.superenhancer_interactome}()}
}
\concept{NOTT_2019}
\keyword{internal}
