% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_loci.R
\name{order_loci}
\alias{order_loci}
\title{Order loci by UCS size, or alphabetically}
\usage{
order_loci(
  dat,
  merged_DT,
  by_UCS_size = FALSE,
  descending = TRUE,
  verbose = FALSE
)
}
\description{
Order loci by UCS size, or alphabetically
}
\examples{
data("merged_DT")
\dontrun{
merged_DT <- echodata::Nalls2019_merged
merged_DT <- order_loci(
    dat = merged_DT,
    merged_DT = merged_DT,
    descending = FALSE
)
}
}
\keyword{internal}
