% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR_process.R
\name{PAINTOR_process}
\alias{PAINTOR_process}
\title{Process PAINTOR files}
\source{
\code{
#### Downloading annotations ####
## Tarball be downloaded manually from here:
https://ucla.box.com/s/x47apvgv51au1rlmuat8m4zdjhcniv2d
## Then decompress the tarball:
tar -xvf Functional_Annotations.tar.gz
}

\code{
#### Installing zen4R ####
## has some system deps that have to be installed beforehand.
sudo apt-get install raptor2-utils
sudo apt-get install rasqal-utils
sudo apt-get install librdf0-dev
}
}
\usage{
PAINTOR_process(
  annot_dir = "./",
  zipfile = "Functional_Annotations",
  sandbox = TRUE,
  title = "PAINTOR",
  token = Sys.getenv("zenodo_token"),
  validate = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{annot_dir}{Directory where \code{Functional_Annotations.tar.gz}
data has been decompressed to.}

\item{sandbox}{Whether to use the Zenodo or Zenodo Sandbox API.}

\item{token}{Zenodo
\href{https://zenodo.org/account/settings/applications/tokens/new/}{
Personal access token}.}

\item{verbose}{Print messages.}
}
\description{
Process PAINTOR annotation/LD-score files and upload them to Zenodo.
}
\keyword{internal}
