% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_overlap_plot.R
\name{peak_overlap_plot}
\alias{peak_overlap_plot}
\title{Plot overlap between some SNP group and various epigenomic data}
\source{
\href{https://science.sciencemag.org/content/366/6469/1134}{
Nott et al. (2019)}
\href{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}{
Corces et al. (2020/bioRxiv)}
}
\usage{
peak_overlap_plot(
  merged_DT,
  snp_filter = "Consensus_SNP==TRUE",
  include.NOTT_2019_peaks = TRUE,
  include.NOTT_2019_enhancers_promoters = TRUE,
  include.NOTT_2019_PLACseq = TRUE,
  include.CORCES_2020_scATACpeaks = TRUE,
  include.CORCES_2020_Cicero_coaccess = TRUE,
  include.CORCES_2020_bulkATACpeaks = TRUE,
  include.CORCES_2020_HiChIP_FitHiChIP_coaccess = TRUE,
  include.CORCES_2020_gene_annotations = TRUE,
  plot_celltype_specificity = TRUE,
  plot_celltype_specificity_genes = FALSE,
  facets_formula = ". ~ Cell_type",
  show_plot = TRUE,
  label_yaxis = TRUE,
  x_strip_angle = 90,
  x_tick_angle = 40,
  drop_empty_cols = F,
  fill_title = paste(snp_filter, "\\nin epigenomic peaks"),
  save_path = F,
  height = 11,
  width = 12,
  subplot_widths = c(1, 0.5),
  verbose = TRUE
)
}
\arguments{
\item{include.NOTT_2019_peaks}{Plot SNP subset overlap with
peaks from cell-type-specific bulk ATAC, H3K27ac, and H3K4me3 assays.}

\item{include.NOTT_2019_enhancers_promoters}{Plot SNP subset overlap with
cell enhancers and promoters.}

\item{include.CORCES_2020_scATACpeaks}{Plot SNP subset overlap with
cell-type-specific scATAC-seq peaks.}

\item{include.CORCES_2020_Cicero_coaccess}{Plot SNP subset overlap with
Cicero coaccessibility peaks (derived from scATACseq).}
}
\description{
Plot overlap between some SNP group and various epigenomic data
}
\examples{
# Consensus SNPs
gg_peaks <- peak_overlap_plot(
    merged_DT = echodata::Nalls2019_merged,
    snp_filter = "Consensus_SNP==T",
    fill_title = "Consensus SNPs in epigenomic peaks"
)
# UCS SNPs
gg_peaks <- peak_overlap_plot(
    merged_DT = echodata::Nalls2019_merged,
    snp_filter = "Support>0",
    fill_title = "UCS SNPs in epigenomic peaks"
)
}
\seealso{
Other summarise: 
\code{\link{CS_bin_plot}()},
\code{\link{CS_counts_plot}()},
\code{\link{get_CS_bins}()},
\code{\link{get_CS_counts}()},
\code{\link{get_SNPgroup_counts}()},
\code{\link{plot_dataset_overlap}()},
\code{\link{results_report}()},
\code{\link{super_summary_plot}()}
}
\concept{summarise}
\keyword{internal}
