% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.get_max_histogram_height.R
\name{PLOT.get_max_histogram_height}
\alias{PLOT.get_max_histogram_height}
\title{PLOT.get_max_histogram_height}
\usage{
PLOT.get_max_histogram_height(gg, round_to = NULL, verbose = TRUE)
}
\description{
PLOT.get_max_histogram_height
}
\keyword{internal}
