% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dataset_overlap.R
\name{plot_dataset_overlap}
\alias{plot_dataset_overlap}
\title{Plot inter-study SNP overlap}
\usage{
plot_dataset_overlap(
  merged_DT,
  snp_filter = "!is.na(SNP)",
  filename = NA,
  formula_str = "~ SNP + Dataset",
  triangle = FALSE,
  proxies = NULL
)
}
\description{
Cross-tabulate SNP overlap (after applying filter)
between each pair of studies.
}
\seealso{
Other summarise: 
\code{\link{CS_bin_plot}()},
\code{\link{CS_counts_plot}()},
\code{\link{get_CS_bins}()},
\code{\link{get_CS_counts}()},
\code{\link{get_SNPgroup_counts}()},
\code{\link{peak_overlap_plot}()},
\code{\link{results_report}()},
\code{\link{super_summary_plot}()}
}
\concept{summarise}
