% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_report.R
\name{results_report}
\alias{results_report}
\title{Give a quick summary report of the fine-mapping results}
\usage{
results_report(merged_dat)
}
\value{
Null
}
\description{
Give a quick summary report of the fine-mapping results
}
\examples{
\dontrun{
results_report(echodata::get_Nalls2019_merged())
}
}
\seealso{
Other summarise: 
\code{\link{CS_bin_plot}()},
\code{\link{CS_counts_plot}()},
\code{\link{peak_overlap_plot}()},
\code{\link{plot_dataset_overlap}()},
\code{\link{super_summary_plot}()}
}
\concept{summarise}
\keyword{internal}
