% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP.construct_reference.R
\name{ROADMAP.construct_reference}
\alias{ROADMAP.construct_reference}
\title{Gather Roadmap annotation metadata}
\usage{
ROADMAP.construct_reference(
  ref_path = system.file("extdata/ROADMAP", "ROADMAP_Epigenomic.js", package =
    "echoannot"),
  keyword_query = NULL
)
}
\arguments{
\item{ref_path}{Where the ROADMAP metadata is stored.}

\item{keyword_query}{Search all columns in the Roadmap annotations metadata
and only query annotations that contain your keywords.
Can provide multiple keywords in list form:
\code{c("placenta","liver","monocytes")}}
}
\description{
Gather Roadmap annotation metadata
}
\examples{
ref <- ROADMAP.construct_reference(keyword_query = c(
    "placenta",
    "liver",
    "monocytes"
))
}
\seealso{
Other ROADMAP: 
\code{\link{ROADMAP.merge_and_process_grl}()},
\code{\link{ROADMAP.query_and_plot}()},
\code{\link{ROADMAP.query}()},
\code{\link{ROADMAP.tabix}()}
}
\concept{ROADMAP}
