% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP.query.R
\name{ROADMAP.query}
\alias{ROADMAP.query}
\title{Query Roadmap by genomic coordinates}
\usage{
ROADMAP.query(
  results_path = file.path(tempdir(), "Roadmap"),
  gr.snp,
  keyword_query = NULL,
  limit_files = NULL,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{results_path}{Where to store query results.}

\item{gr.snp}{\code{\link[GenomicRanges]{GRanges}} object of 
SNPs to query Roadmap with.}

\item{limit_files}{Limit the number of annotation files queried 
(for faster testing).}

\item{nThread}{Number of threads to parallelise queries over.}
}
\description{
Query Roadmap by genomic coordinates
}
\examples{
\dontrun{
grl.roadmap <- ROADMAP.query(
    gr.snp = echodata::BST1,
    keyword_query = "placenta")
}
}
\seealso{
Other ROADMAP: 
\code{\link{ROADMAP.construct_reference}()},
\code{\link{ROADMAP.merge_and_process_grl}()},
\code{\link{ROADMAP.query_and_plot}()},
\code{\link{ROADMAP.tabix}()}
}
\concept{ROADMAP}
