% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP.tabix.R
\name{ROADMAP.tabix}
\alias{ROADMAP.tabix}
\title{Query Roadmap API}
\usage{
ROADMAP.tabix(
  results_path = tempfile(fileext = "ROADMAP_query.csv.gz"),
  chrom,
  min_pos,
  max_pos,
  eid,
  convert_to_granges = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{results_path}{Where to store query results.}

\item{chrom}{Chromosome to query}

\item{min_pos}{Minimum genomic position}

\item{max_pos}{Maximum genomic position}

\item{eid}{Roadmap annotation ID}

\item{convert_to_granges}{Whether to return query
as a \code{data.frame} or \code{\link[GenomicRanges]{GRanges}}.}
}
\description{
Query Roadmap epigenomic annotations (chromatin marks)
using a range of genomic coordinates.
}
\details{
\href{https://egg2.wustl.edu/roadmap/data/byFileType/chromhmmSegmentations/ChmmModels/coreMarks/jointModel/final/}{
ROADMAP file locations.}
}
\examples{
\dontrun{
BST1 <- echodata::BST1
dat <- ROADMAP.tabix(chrom=BST1$CHR[1],
                     min_pos = min(BST1$POS), 
                     max_pos = max(BST1$POS), 
                     eid = "E099")
}

}
\seealso{
Other ROADMAP: 
\code{\link{ROADMAP.construct_reference}()},
\code{\link{ROADMAP.merge_and_process_grl}()},
\code{\link{ROADMAP.query_and_plot}()},
\code{\link{ROADMAP.query}()}
}
\concept{ROADMAP}
\keyword{internal}
