% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP.track_plot.R
\name{ROADMAP.track_plot}
\alias{ROADMAP.track_plot}
\title{Plot Roadmap query}
\usage{
ROADMAP.track_plot(
  grl.roadmap.filt,
  gr.snp = NULL,
  geom = "density",
  adjust = 0.2,
  show_plot = TRUE,
  as.ggplot = TRUE
)
}
\arguments{
\item{grl.roadmap.filt}{Roadmap query results.}

\item{gr.snp}{Optionally, can include an extra \code{\link[GenomicRanges]{GRanges}} object
to ensure the plot does not extend beyond certain coordinates.}

\item{geom}{The type of plot to create.
Options include "density" and "histogram".}

\item{adjust}{The granularity of the peaks.}

\item{show_plot}{Whether to print the plot.}
}
\description{
Plot Roadmap query
}
\examples{
\dontrun{
gr.snp <- dt_to_granges(echodata::BST1)
grl.roadmap <- ROADMAP.query(
    gr.snp = gr.snp,
    keyword_query = "monocyte"
)
grl.roadmap.filt <- ROADMAP.merge_and_process_grl(
    grl.roadmap = grl.roadmap,
    gr.snp = gr.snp
)
track.roadmap <- ROADMAP.track_plot(grl.roadmap.filt,
    gr.snp = gr.snp
)
}
}
\keyword{internal}
