% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP_merge_and_process.R
\name{ROADMAP_merge_and_process}
\alias{ROADMAP_merge_and_process}
\title{Standardize Roadmap query}
\usage{
ROADMAP_merge_and_process(
  grl.roadmap,
  gr.snp,
  n_top = NULL,
  minoverlap = 1,
  verbose = TRUE
)
}
\arguments{
\item{grl.roadmap}{Roadmap query results.}

\item{n_top}{The number of top annotation sources (e.g. tissues)
 to include, sorted by greatest number of rows
(i.e. the number of genomic ranges within the window).}

\item{minoverlap}{
    A single non-negative integer.

    Only ranges with a minimum of \code{minoverlap} overlapping positions
    are considered to be overlapping.

    When \code{type} is \code{"any"}, at least one of \code{maxgap} and
    \code{minoverlap} must be set to its default value.
  }

\item{verbose}{Print messages.}
}
\description{
Standardize and filter Roadmap query results.
}
\seealso{
Other ROADMAP: 
\code{\link{ROADMAP_construct_reference}()},
\code{\link{ROADMAP_query}()},
\code{\link{ROADMAP_tabix}()}
}
\concept{ROADMAP}
\keyword{internal}
