% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP_merge_and_process_grl.R
\name{ROADMAP_merge_and_process_grl}
\alias{ROADMAP_merge_and_process_grl}
\title{Standardize Roadmap query}
\usage{
ROADMAP_merge_and_process_grl(
  grl.roadmap,
  gr.snp,
  n_top_tissues = 5,
  sep = " ",
  verbose = TRUE
)
}
\arguments{
\item{grl.roadmap}{Roadmap query results}

\item{n_top_tissues}{The number of top tissues to include,
sorted by greatest number of rows
(i.e. the number of genomic ranges within the window).}

\item{verbose}{Print messages.}
}
\description{
Standardize Roadmap query
}
\seealso{
Other ROADMAP: 
\code{\link{ROADMAP_construct_reference}()},
\code{\link{ROADMAP_query}()},
\code{\link{ROADMAP_tabix}()}
}
\concept{ROADMAP}
\keyword{internal}
