% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP_plot.R
\name{ROADMAP_plot}
\alias{ROADMAP_plot}
\title{Plot annotation peaks: ROADMAP}
\usage{
ROADMAP_plot(
  dat,
  roadmap_query,
  lib_name = "Roadmap.ChromatinMarks_CellTypes",
  locus_dir = tempdir(),
  n_top = 5,
  force_new = FALSE,
  show_plot = FALSE,
  conda_env = "echoR",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{dat}{\link[data.table]{data.table} with at least the following columns:
\describe{
\item{SNP}{SNP RSID}
\item{CHR}{chromosome}
\item{POS}{position}
}}

\item{n_top}{Number of top annotations to be plotted
(passed to \link[echoannot]{ROADMAP_merge_and_process_grl}).}

\item{force_new}{Download and prepare a new query
even if the file already exists locally (Default: \code{FALSE}).}

\item{show_plot}{Print the plot.}

\item{conda_env}{Conda environment to search for tabix in.}

\item{nThread}{Number of threads to parallelise queries over.}
}
\value{
A named list containing:
\itemize{
\item{"data"}{\code{GRanges} object within the queried coordinates.}
\item{"plot"}{\code{ggbio} plot.}
}
}
\description{
Import annotation from \emph{ROADMAP}, filter to only those within the range
of \code{dat}, and then plot the peaks.
If the annotation has already been downloaded previously, it will be reused.
}
\examples{
roadmap_out <- echoannot::ROADMAP_plot(
    dat = echodata::BST1[1:1000, ],
    roadmap_query = "placenta"
)
}
