% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP_tabix.R
\name{ROADMAP_tabix}
\alias{ROADMAP_tabix}
\title{Query Roadmap API}
\usage{
ROADMAP_tabix(
  results_path = tempfile(fileext = "ROADMAP_query.csv.gz"),
  chrom,
  min_pos,
  max_pos,
  eid,
  convert_to_granges = TRUE,
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\arguments{
\item{results_path}{Where to store query results.}

\item{chrom}{Chromosome to query}

\item{min_pos}{Minimum genomic position}

\item{max_pos}{Maximum genomic position}

\item{eid}{Roadmap annotation ID}

\item{convert_to_granges}{Whether to return query
as a \code{data.frame} or \code{\link[GenomicRanges]{GRanges}}.}
}
\description{
Query Roadmap epigenomic annotations (chromatin marks)
using a range of genomic coordinates.
}
\details{
\href{https://egg2.wustl.edu/roadmap/data/byFileType/chromhmmSegmentations/ChmmModels/coreMarks/jointModel/final/}{
ROADMAP file locations.}
}
\examples{
\dontrun{
BST1 <- echodata::BST1
dat <- ROADMAP_tabix(
    chrom = BST1$CHR[1],
    min_pos = min(BST1$POS),
    max_pos = max(BST1$POS),
    eid = "E099"
)
}

}
\seealso{
Other ROADMAP: 
\code{\link{ROADMAP_construct_reference}()},
\code{\link{ROADMAP_merge_and_process_grl}()},
\code{\link{ROADMAP_query}()}
}
\concept{ROADMAP}
\keyword{internal}
