% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_genome.R
\name{select_genome}
\alias{select_genome}
\title{Select genome build}
\usage{
select_genome(genome_build, verbose = TRUE)
}
\arguments{
\item{genome_build}{Genome build to use.}

\item{verbose}{Print messages.}
}
\value{
Named list
}
\description{
Select and load a valid genome build and its relevant  Bioc data packages.
}
\keyword{internal}
