% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps_by_mutation_type.R
\name{snps_by_mutation_type}
\alias{snps_by_mutation_type}
\title{Return only the missense SNPs}
\source{
\href{https://bioconductor.org/packages/release/bioc/html/biomaRt.html}{
biomaRt}
}
\usage{
snps_by_mutation_type(merged_results, mutation_type = "missense_variant")
}
\description{
Return only the missense SNPs
}
\seealso{
Other annotate: 
\code{\link{annotate_missense}()},
\code{\link{biomart_geneInfo}()},
\code{\link{biomart_snp_info}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{haplor_epigenetics_enrichment}()},
\code{\link{haplor_epigenetics_summary}()},
\code{\link{haplor_haploreg}()},
\code{\link{haplor_regulomedb}()},
\code{\link{plot_missense}()}
}
\concept{annotate}
\keyword{internal}
