% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_summary_plot.R
\name{super_summary_plot}
\alias{super_summary_plot}
\title{Merge all summary plots into one super plot}
\usage{
super_summary_plot(
  merged_DT,
  snp_filter = "Consensus_SNP==TRUE",
  coloc_results = NULL,
  plot_missense = TRUE,
  show_plot = TRUE,
  save_plot = FALSE,
  height = 15,
  width = 13,
  dpi = 500
)
}
\description{
Merge all summary plots into one super plot
}
\seealso{
Other summarise: 
\code{\link{CS_bin_plot}()},
\code{\link{CS_counts_plot}()},
\code{\link{get_CS_bins}()},
\code{\link{get_CS_counts}()},
\code{\link{get_SNPgroup_counts}()},
\code{\link{peak_overlap_plot}()},
\code{\link{plot_dataset_overlap}()},
\code{\link{results_report}()}
}
\concept{summarise}
