% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_to_ggplot_list.R
\name{tracks_to_ggplot_list}
\alias{tracks_to_ggplot_list}
\title{Convert ggbio tracks plot to ggplot2 list}
\usage{
tracks_to_ggplot_list(trks)
}
\arguments{
\item{trks}{ggbio tracks plot.}
}
\description{
Convert \link[ggbio]{tracks} plot to a named list of 
\link[ggplot2]{ggplot} objects.
}
\keyword{internal}
