% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cols.R
\name{update_cols}
\alias{update_cols}
\title{Update CS cols}
\usage{
update_cols(finemap_dat)
}
\description{
Convert old column format: ".Credible_Set" => ".CS"
}
\examples{
\dontrun{
finemap_DT <- update_cols(finemap_dat = echodata::BST1)
}
}
\keyword{internal}
