% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.download_and_standardize.R
\name{XGR.download_and_standardize}
\alias{XGR.download_and_standardize}
\title{Download, standardize, and merge XGR annotations}
\usage{
XGR.download_and_standardize(
  lib.selections = c("ENCODE_TFBS_ClusteredV3_CellTypes", "TFBS_Conserved",
    "Uniform_TFBS"),
  as_grangesList = FALSE,
  finemap_dat,
  nThread = 1
)
}
\arguments{
\item{lib.selections}{Which XGR annotations to check overlap with.
For full list of libraries see
\href{http://xgr.r-forge.r-project.org/#annotations-at-the-genomic-region-level}{
 here.}}

\item{as_grangesList}{Return as a \code{GRangesList},
instead of a single merged \code{GRanges} object.}

\item{finemap_dat}{Fine-mapping results.}

\item{nThread}{Number of cores to parallelise across.}
}
\value{
GRangesList
}
\description{
Merges a list of XGR annotations into a single GRanges object
}
\examples{
\dontrun{
gr.lib <- XGR.download_and_standardize(
    lib.selections = c("ENCODE_DNaseI_ClusteredV3_CellTypes"),
    finemap_dat = echodata::BST1)
}
}
\seealso{
Other XGR: 
\code{\link{XGR.enrichment_bootstrap}()},
\code{\link{XGR.enrichment_plot}()},
\code{\link{XGR.enrichment}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.parse_metadata}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()},
\code{\link{XGR.sep_handler}()},
\code{\link{granges_to_bed}()}
}
\concept{XGR}
