% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.parse_metadata.R
\name{XGR.parse_metadata}
\alias{XGR.parse_metadata}
\title{XGR.parse_metadata}
\usage{
XGR.parse_metadata(gr.lib, lib.name = NA)
}
\description{
XGR.parse_metadata
}
\seealso{
Other XGR: 
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.enrichment_bootstrap}()},
\code{\link{XGR.enrichment_plot}()},
\code{\link{XGR.enrichment}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()},
\code{\link{XGR.sep_handler}()},
\code{\link{granges_to_bed}()}
}
\concept{XGR}
\keyword{internal}
