% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.prepare_foreground_background.R
\name{XGR.prepare_foreground_background}
\alias{XGR.prepare_foreground_background}
\title{Prepare SNP sets for enrichment}
\usage{
XGR.prepare_foreground_background(
  subset_DT,
  foreground_filter = "Support>0",
  background_filter = NULL,
  fg_sample_size = NULL,
  bg_sample_size = NULL,
  verbose = TRUE
)
}
\arguments{
\item{subset_DT}{Data.frame with at least the following columns:
\describe{
\item{SNP}{SNP RSID}
\item{CHR}{chromosome}
\item{POS}{position}
}}

\item{foreground_filter}{Specify foreground by filtering SNPs
in \code{subset_DT}.
Write filter as a string (or \code{NULL} to include all SNPs).}

\item{background_filter}{Specify background by filtering SNPs
in \code{subset_DT}.
Write filter as a string (or \code{NULL} to include all SNPs).}
}
\description{
Prepare custom foreground and background SNPs sets for enrichment
tests with XGR annotations.
}
\examples{
\dontrun{
fg_bg <- XGR.prepare_foreground_background(
    subset_DT = echodata::get_Nalls2019_merged(),
    foreground_filter = "Consensus_SNP==TRUE",
    background_filter = "leadSNP==TRUE"
)
}
}
\seealso{
Other XGR: 
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.enrichment_bootstrap}()},
\code{\link{XGR.enrichment_plot}()},
\code{\link{XGR.enrichment}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.parse_metadata}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.sep_handler}()},
\code{\link{granges_to_bed}()}
}
\concept{XGR}
\keyword{internal}
