% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_enrichment.R
\name{XGR_enrichment}
\alias{XGR_enrichment}
\title{XGR enrichment}
\usage{
XGR_enrichment(
  gr,
  merged_dat,
  foreground_filter = "Consensus_SNP==TRUE",
  background_filter = NULL,
  grouping_vars = c("Study", "Assay", "Cell_type"),
  fg_sample_size = NULL,
  bg_sample_size = NULL,
  background.annotatable.only = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{gr}{Annotations to test for enrichment with.}

\item{merged_dat}{SNP-level fine-mapping results to test for enrichment with.}

\item{foreground_filter}{Filter to apply to foreground (target SNPs).}

\item{background_filter}{Filter to apply to background (non-target SNPs).}

\item{grouping_vars}{Columns in \code{merged_dat} to group by when conducting
enrichment tests.}

\item{fg_sample_size}{Foreground sample size.}

\item{bg_sample_size}{Background sample size.}

\item{background.annotatable.only}{For background SNPs,
only use SNPs that overlap with some annotation in \code{gr}.
 This means that missing annotations (\code{NA}) will not be considered.}

\item{verbose}{Print messages.}
}
\description{
Run SNP-level enrichment test with \link[XGR]{xGRviaGenomicAnno}.
}
\examples{
\dontrun{
gr.merged <- echoannot::merge_celltype_specific_epigenomics()
enrich.lead <- XGR_enrichment(
    gr = gr.merged,
    merged_dat = echodata::get_Nalls2019_merged(),
    foreground_filter = "leadSNP==TRUE",
    grouping_vars = c("Study", "Cell_type", "Assay")
)
}
}
\seealso{
Other XGR: 
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_query}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_example}}
}
\concept{XGR}
