% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_enrichment_bootstrap.R
\name{XGR_enrichment_bootstrap}
\alias{XGR_enrichment_bootstrap}
\title{XGR enrichment (bootstrapped)}
\usage{
XGR_enrichment_bootstrap(
  gr,
  merged_dat,
  snp_groups = c("Random", "GWAS lead", "UCS (-PolyFun)", "UCS", "Consensus (-PolyFun)",
    "Consensus"),
  background_filter = NULL,
  grouping_vars = c("Study", "Assay", "Cell_type"),
  iterations = 1000,
  fg_sample_size = 20,
  bg_sample_size = NULL,
  bootstrap = TRUE,
  save_path = tempfile(fileext = "XGR_enrich_boot_res.csv.gz"),
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{gr}{Annotations to test for enrichment with.}

\item{merged_dat}{SNP-level fine-mapping results to test for enrichment with.}

\item{snp_groups}{Which SNP groups to repeat enrichment tests for separately.}

\item{background_filter}{Filter to apply to background (non-target SNPs).}

\item{grouping_vars}{Columns in \code{merged_dat} to group by when conducting
enrichment tests.}

\item{iterations}{Number of bootstrapping iterations.}

\item{fg_sample_size}{Foreground sample size.}

\item{bg_sample_size}{Background sample size.}

\item{bootstrap}{Whether to use bootstrapping.}

\item{save_path}{File path to save results to.}

\item{nThread}{Number of threads to parallelise bootstrapping over.}

\item{verbose}{Print messages.}
}
\description{
Perform annotation enrichment tests using iterative bootstrapping procedure.
}
\examples{
\dontrun{
gr.merged <- echoannot::merge_celltype_specific_epigenomics()
enrich_res <- XGR_enrichment_bootstrap(
    gr = gr.merged,
    merged_dat = echodata::get_Nalls2019_merged()
)
}
}
\seealso{
Other XGR: 
\code{\link{XGR_download_and_standardize}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_plot_peaks}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_query}}
}
\concept{XGR}
