% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_enrichment_plot.R
\name{XGR_enrichment_plot}
\alias{XGR_enrichment_plot}
\title{Plot enrichment results}
\usage{
XGR_enrichment_plot(
  enrich_res,
  title = NULL,
  subtitle = NULL,
  facet_formula = NULL,
  line_formula = "y ~ x",
  line_method = "lm",
  line_span = 1,
  FDR_thresh = 1,
  plot_type = "bar",
  shape_var = "Cell_type",
  facet_scales = "free",
  show_plot = TRUE,
  save_plot = FALSE,
  height = 5,
  width = 5,
  dpi = 300
)
}
\arguments{
\item{height}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{width}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}
}
\description{
Plot enrichment results from XGR annotations.
}
\examples{
\dontrun{
root <- file.path(
    "/sc/arion/projects/pd-omics/brian",
    "Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide"
)
### merged enrichment results
enrich_res <- data.table::fread(
    file.path(
        root,
        "XGR/celltypespecific_epigenomics.SNP_groups.csv.gz"
    )
)
enrich_res <- data.table::fread(
    file.path(
        root,
        "XGR/celltypespecific_epigenomics.snp_groups.csv.gz"
    )
)
enrich_boot <- data.table::fread(
    file.path(
        root,
        "XGR/celltypespecific_epigenomics.snp_groups.permute.csv.gz"
    )
)
enrich_assay <- data.table::fread(
    file.path(
        root,
        "XGR/celltypespecific_epigenomics.snp_groups.assay.csv.gz"
    )
)

# Merged volcano plot
enrich_res <- subset(enrich_res, SNP_Group != "Consensus (-PolyFun)") \%>\%
    dplyr::rename(SNP_group = SNP_Group)
gp <- XGR_enrichment_plot(
    enrich_res = subset(enrich_res, !Assay \%in\% c("HiChIP_FitHiChIP", "PLAC")),
    title = "Enrichment: Cell-type-specific epigenomics",
    plot_type = "point",
    save_plot = file.path(
        root, "XGR/celltypespecific_epigenomics.enrich_volcano.png"
    ),
    height = 6, width = 8, shape_var = "Assay"
)
## Merged bar plot
gp <- XGR_enrichment_plot(
    enrich_res = enrich_res,
    plot_type = "bar",
    facet_formula = ".~Assay",
    FDR_thresh = .05
)
# Merged volcano plot (permuted)
gp <- XGR_enrichment_plot(
    enrich_res = enrich.scATAC.permute,
    title = "Permuted enrichment: Cell-type-specific peaks and elements",
    plot_type = "point"
)
}
}
\seealso{
Other XGR: 
\code{\link{XGR_download_and_standardize}()},
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_plot_peaks}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_query}}
}
\concept{XGR}
\keyword{internal}
