% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xgr_example}
\alias{xgr_example}
\title{Example XGR query}
\format{
An object of class \code{GRanges} of length 8490.
}
\source{
\code{
xgr_example <- echoannot::XGR_query(
    c("ENCODE_DNaseI_ClusteredV3_CellTypes"),
    dat = echodata::BST1)
}
usethis::use_data(xgr_example, overwrite = TRUE)
}
\usage{
data("xgr_example")
}
\description{
Example XGR query results from \pkg{XGR} via
\link[echoannot]{XGR_query} using the
"ENCODE_DNaseI_ClusteredV3_CellTypes" dataset.
}
\seealso{
Other XGR: 
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_query}()},
\code{\link{XGR_sep_handler}()}
}
\concept{XGR}
\keyword{datasets}
