% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_filter_assays.R
\name{XGR_filter_assays}
\alias{XGR_filter_assays}
\title{Filter assays}
\usage{
XGR_filter_assays(gr.lib, n_top_assays = 5)
}
\arguments{
\item{gr.lib}{\code{GRanges} object of annotations.}

\item{n_top_assays}{Number of top assays to include per library.}
}
\description{
Identify the assays with the most annotations in the locus.
Then only keep these assays.
}
\examples{
gr.filt <- echoannot::XGR_filter_assays(gr.lib = echoannot::xgr_query)
}
\seealso{
Other XGR: 
\code{\link{XGR_download_and_standardize}()},
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_plot_peaks}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_query}}
}
\concept{XGR}
