% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_filter_sources.R
\name{XGR_filter_sources}
\alias{XGR_filter_sources}
\title{Filter sources}
\usage{
XGR_filter_sources(gr.lib, n_top_sources = 1)
}
\arguments{
\item{gr.lib}{Results from \link[echoannot]{XGR_query}.}

\item{n_top_sources}{Number of top sources to include per library.}
}
\description{
Identify the sources with the most annotations in the locus.
Then only keep these sources.
}
\examples{
gr.filt <- echoannot::XGR_filter_sources(gr.lib = echoannot::xgr_example)
}
\seealso{
Other XGR: 
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_query}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_example}}
}
\concept{XGR}
