% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_import_annotations.R
\name{XGR_import_annotations}
\alias{XGR_import_annotations}
\title{Download XGR annotations}
\usage{
XGR_import_annotations(
  gr.snp,
  anno_data_path = file.path("annotations", paste0("XGR_", lib.name, ".rds")),
  lib.name,
  save_xgr = TRUE,
  annot_overlap_threshold = 5
)
}
\description{
Download XGR annotations
}
\seealso{
Other XGR: 
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_query}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_example}}
}
\concept{XGR}
\keyword{internal}
