% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_iterate_enrichment.R
\name{XGR_iterate_enrichment}
\alias{XGR_iterate_enrichment}
\title{Conduct enrichment tests for each annotation}
\usage{
XGR_iterate_enrichment(
  dat,
  foreground_filter = "Consensus_SNP",
  background_filter = "leadSNP",
  lib.selections = c("ENCODE_TFBS_ClusteredV3_CellTypes",
    "ENCODE_DNaseI_ClusteredV3_CellTypes", "Broad_Histone", "FANTOM5_Enhancer",
    "Segment_Combined_Gm12878", "TFBS_Conserved", "ReMap_PublicAndEncode_TFBS",
    "Blueprint_VenousBlood_Histone", "Blueprint_DNaseI", "FANTOM5_CAT_Cell",
    "FANTOM5_CAT_MESH", "GWAScatalog_alltraits"),
  save_path = FALSE,
  nThread = 1
)
}
\arguments{
\item{dat}{Data.frame with at least the following columns:
\describe{
\item{SNP}{SNP RSID}
\item{CHR}{chromosome}
\item{POS}{position}
}}

\item{foreground_filter}{Specify foreground by filtering SNPs
in \code{dat}.
Write filter as a string (or \code{NULL} to include all SNPs).}

\item{background_filter}{Specify background by filtering SNPs
in \code{dat}.
Write filter as a string (or \code{NULL} to include all SNPs).}
}
\description{
XGR uses a binomial enrichment tests for each annotation.
}
\details{
\href{https://www.rdocumentation.org/packages/XGR/versions/1.1.5/topics/xDefineGenomicAnno}{
Description of all datasets}
}
\examples{
\dontrun{
enrich_res <- XGR_iterate_enrichment(
    dat = echodata::get_Nalls2019_merged(),
    foreground_filter = "Consensus_SNP",
    background_filter = "leadSNP",
    lib.selections = c("ENCODE_TFBS_ClusteredV3_CellTypes")
)
}
}
\seealso{
Other XGR: 
\code{\link{XGR_download_and_standardize}()},
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_plot_peaks}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_query}}
}
\concept{XGR}
\keyword{internal}
