% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_merge_and_process.R
\name{XGR_merge_and_process}
\alias{XGR_merge_and_process}
\title{Standardize XGR annotations}
\usage{
XGR_merge_and_process(grl.xgr, lib, n_top_sources = 10)
}
\arguments{
\item{grl.xgr}{\link[GenomicRanges]{GenomicRangesList} of XGR queries.}
}
\description{
Parses the metadata and adds it as columns,
and then merges the results into a single
\code{\link[GenomicRanges]{GenomicRangesList}}
}
\seealso{
Other XGR: 
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_query}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_example}}
}
\concept{XGR}
\keyword{internal}
