% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_parse_metadata.R
\name{XGR_parse_metadata}
\alias{XGR_parse_metadata}
\title{XGR_parse_metadata}
\usage{
XGR_parse_metadata(gr.lib, lib.name = NA)
}
\description{
XGR_parse_metadata
}
\seealso{
Other XGR: 
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_query}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_example}}
}
\concept{XGR}
\keyword{internal}
