% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_plot.R
\name{XGR_plot}
\alias{XGR_plot}
\title{Plot annotation peaks: XGR}
\usage{
XGR_plot(
  dat,
  lib_name = "ENCODE_TFBS_ClusteredV3_CellTypes",
  locus_dir = tempdir(),
  palette = c("Spectral", "BrBG", "PiYG", "PuOr"),
  fill_var = "Assay",
  facet_var = "Source",
  geom = "density",
  n_top = 5,
  force_new = FALSE,
  show_plot = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{dat}{\link[data.table]{data.table} with at least the following columns:
\describe{
\item{SNP}{SNP RSID}
\item{CHR}{chromosome}
\item{POS}{position}
}}

\item{fill_var}{Fill variable.}

\item{facet_var}{Row facet variable.}

\item{geom}{Plot type ("density", or "histogram").}

\item{n_top}{Number of top annotations to be plotted
(passed to \link[echoannot]{XGR_filter_sources} and then
\link[echoannot]{XGR_filter_assays}).}

\item{force_new}{Download and prepare a new query
even if the file already exists locally (Default: \code{FALSE}).}

\item{show_plot}{Print the plot.}
}
\value{
List with the "data" and the "plot".
}
\description{
Import annotation from \emph{XGR}, filter to only those within the range of
\code{dat}, and then plot the peaks.
If the annotation has already been downloaded previously, it will be reused.
}
\examples{
xgr_out <- echoannot::XGR_plot(dat = echodata::BST1[1:1000, ])
}
