% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_plot_enrichment.R
\name{XGR_plot_enrichment}
\alias{XGR_plot_enrichment}
\title{Plot XGR enrichment}
\usage{
XGR_plot_enrichment(
  enrich_res,
  adjp_thresh = 0.05,
  top_annotations = NULL,
  show_plot = TRUE
)
}
\description{
Plot XGR enrichment.
}
\examples{
\dontrun{
enrich_res <- XGR_iterate_enrichment(
    dat = echodata::get_Nalls2019_merged(),
    foreground_filter = "Consensus_SNP",
    background_filter = "leadSNP",
    lib.selections = c("ENCODE_TFBS_ClusteredV3_CellTypes")
)
XGR_plot_enrichment(enrich_res)
}
}
\seealso{
Other XGR: 
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_query}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_example}}
}
\concept{XGR}
\keyword{internal}
