% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_plot_peaks.R
\name{XGR_plot_peaks}
\alias{XGR_plot_peaks}
\title{Plot XGR peaks}
\usage{
XGR_plot_peaks(
  gr.lib,
  dat,
  fill_var = "Assay",
  facet_var = "Source",
  geom = "density",
  locus = NULL,
  adjust = 0.2,
  show_plot = TRUE,
  show.legend = TRUE,
  as.ggplot = TRUE,
  trim_xlims = FALSE
)
}
\arguments{
\item{gr.lib}{\code{GRanges} object of annotations.}

\item{dat}{Data.frame with at least the following columns:
\describe{
\item{SNP}{SNP RSID}
\item{CHR}{chromosome}
\item{POS}{position}
}}

\item{fill_var}{Fill variable.}

\item{facet_var}{Row facet variable.}

\item{geom}{Plot type ("density", or "histogram").}

\item{locus}{[Optional] Locus name.}

\item{adjust}{The granularity of the peaks.}

\item{show_plot}{Print the plot.}
}
\value{
\code{ggbio} track plot.
}
\description{
Plots the distribution of annotations across a genomic region (x-axis).
}
\examples{
#### Import example query ####
gr.lib <- echoannot::xgr_query

#### Filter query ####
gr.filt <- echoannot::XGR_filter_sources(gr.lib = gr.lib)
gr.filt <- echoannot::XGR_filter_assays(gr.lib = gr.filt)

#### Plot query #####
XGR_track <- echoannot::XGR_plot_peaks(
    gr.lib = gr.filt,
    dat = echodata::BST1,
    fill_var = "Assay",
    facet_var = "Source"
)
}
\seealso{
Other XGR: 
\code{\link{XGR_download_and_standardize}()},
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_query}}
}
\concept{XGR}
