% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_prepare_foreground_background.R
\name{XGR_prepare_foreground_background}
\alias{XGR_prepare_foreground_background}
\title{Prepare SNP sets for enrichment}
\usage{
XGR_prepare_foreground_background(
  dat,
  foreground_filter = "Support>0",
  background_filter = NULL,
  fg_sample_size = NULL,
  bg_sample_size = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Data.frame with at least the following columns:
\describe{
\item{SNP}{SNP RSID}
\item{CHR}{chromosome}
\item{POS}{position}
}}

\item{foreground_filter}{Specify foreground by filtering SNPs
in \code{dat}.
Write filter as a string (or \code{NULL} to include all SNPs).}

\item{background_filter}{Specify background by filtering SNPs
in \code{dat}.
Write filter as a string (or \code{NULL} to include all SNPs).}
}
\description{
Prepare custom foreground and background SNPs sets for enrichment
tests with XGR annotations.
}
\examples{
\dontrun{
fg_bg <- XGR_prepare_foreground_background(
    dat = echodata::get_Nalls2019_merged(),
    foreground_filter = "Consensus_SNP==TRUE",
    background_filter = "leadSNP==TRUE"
)
}
}
\seealso{
Other XGR: 
\code{\link{XGR_download_and_standardize}()},
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_plot_peaks}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_query}}
}
\concept{XGR}
\keyword{internal}
