% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_query.R
\name{XGR_query}
\alias{XGR_query}
\title{Download, standardize, and merge XGR annotations}
\usage{
XGR_query(
  lib.selections = c("ENCODE_TFBS_ClusteredV3_CellTypes", "TFBS_Conserved",
    "Uniform_TFBS"),
  as_grangesList = FALSE,
  dat = NULL,
  n_top = NULL,
  nThread = 1
)
}
\arguments{
\item{lib.selections}{Which XGR annotations to check overlap with.
For full list of libraries see
\href{http://XGR_r-forge.r-project.org/#annotations-at-the-genomic-region-level}{
 here.}
 Passed to the \code{RData.customised} argument in \link[XGR]{xRDataLoader}.}

\item{as_grangesList}{Return as a \link[GenomicRanges]{GRangesList},
instead of a single merged \link[GenomicRanges]{GRanges} object.}

\item{dat}{data.table of genomic coordinates to query with.
Set as \code{NULL} to return genome-wide data.}

\item{n_top}{Filter to only the top N annotations 
that have the greatest amount of overlap with the genomic coordinates of
\code{dat}.}

\item{nThread}{Number of cores to parallelise across.}
}
\value{
\link[GenomicRanges]{GRangesList}
}
\description{
Merges a list of XGR annotations into a single
 \link[GenomicRanges]{GRangesList} 
 (or merged \link[GenomicRanges]{GRanges}) object.
}
\examples{
gr.lib <- echoannot::XGR_query(
    lib.selections = c("ENCODE_DNaseI_ClusteredV3_CellTypes"),
    dat = echodata::BST1, 
    n_top = 1)
}
\seealso{
Other XGR: 
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{XGR_sep_handler}()},
\code{\link{xgr_example}}
}
\concept{XGR}
