% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_sep_handler.R
\name{XGR_sep_handler}
\alias{XGR_sep_handler}
\title{XGR_sep_handler}
\usage{
XGR_sep_handler(lib.name)
}
\description{
XGR_sep_handler
}
\seealso{
Other XGR: 
\code{\link{XGR_download_and_standardize}()},
\code{\link{XGR_enrichment_bootstrap}()},
\code{\link{XGR_enrichment_plot}()},
\code{\link{XGR_enrichment}()},
\code{\link{XGR_filter_assays}()},
\code{\link{XGR_filter_sources}()},
\code{\link{XGR_import_annotations}()},
\code{\link{XGR_iterate_enrichment}()},
\code{\link{XGR_iterate_overlap}()},
\code{\link{XGR_merge_and_process}()},
\code{\link{XGR_parse_metadata}()},
\code{\link{XGR_plot_enrichment}()},
\code{\link{XGR_plot_peaks}()},
\code{\link{XGR_prepare_foreground_background}()},
\code{\link{xgr_query}}
}
\concept{XGR}
\keyword{internal}
