% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_list.R
\name{zenodo_list}
\alias{zenodo_list}
\title{List Zenodo files}
\source{
\href{https://cran.r-project.org/web/packages/zen4R/vignettes/zen4R.html}{
zen4R vignette}
\href{https://rpubs.com/antaldaniel/zenodo-sandbox-setup}{
zen4R vignette}
}
\usage{
zenodo_list(
  token,
  concept_doi = "10.5281/zenodo.7062237",
  sandbox = FALSE,
  as_datatable = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{token}{Zenodo
\href{https://zenodo.org/account/settings/applications/tokens/new/}{
Personal access token}.}

\item{concept_doi}{Concept DOI of the Zenodo record.}

\item{sandbox}{Whether to use the Zenodo or Zenodo Sandbox API.}

\item{as_datatable}{Return file info as a \link[data.table]{data.table}.}

\item{verbose}{Print messages.}
}
\value{
Nested list of Zenodo file info.
}
\description{
List all files available in a given Zenodo record.
Requires already having a Zenodo account set up and
}
\keyword{internal}
