% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_specificity.R
\name{cell_type_specificity}
\alias{cell_type_specificity}
\title{Get cell-type-specifity score for each cell type}
\usage{
cell_type_specificity(
  plot_dat,
  merged_DT,
  plot_specificity = TRUE,
  min_count = NULL,
  top_celltype_only = FALSE,
  label_yaxis = TRUE,
  y_lab = NULL,
  show_genes = FALSE,
  x_strip_angle = 40,
  show_plot = TRUE
)
}
\arguments{
\item{plot_specificity}{Plot cell-type specificity scores 
for locus instead of raw assay overlap counts.}
}
\description{
Aggregate SNP overlap across various epigenomic datasets
and then identify the number of SNPs overlapping by each cell type
}
\keyword{internal}
