% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS_bin_plot.R
\name{CS_bin_plot}
\alias{CS_bin_plot}
\title{Plot CS bin counts}
\usage{
CS_bin_plot(merged_DT, show_plot = TRUE)
}
\description{
Plot CS bin counts
}
\examples{
bin_plot <- CS_bin_plot(merged_DT = echodata::get_Nalls2019_merged())
}
\seealso{
Other summarise: 
\code{\link{CS_counts_plot}()},
\code{\link{get_CS_bins}()},
\code{\link{get_CS_counts}()},
\code{\link{get_SNPgroup_counts}()},
\code{\link{peak_overlap_plot}()},
\code{\link{plot_dataset_overlap}()},
\code{\link{results_report}()},
\code{\link{super_summary_plot}()}
}
\concept{summarise}
