% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS_counts_plot.R
\name{CS_counts_plot}
\alias{CS_counts_plot}
\title{Bar plot of tool-specific CS sizes}
\usage{
CS_counts_plot(
  merged_DT,
  show_numbers = TRUE,
  ylabel = "Locus",
  legend_nrow = 3,
  label_yaxis = TRUE,
  top_CS_only = FALSE,
  show_plot = TRUE
)
}
\arguments{
\item{merged_DT}{Merged fine-mapping results data from
\link[echolocatoR]{finemap_loci}.}

\item{show_numbers}{Print numbers on top of bars.}

\item{ylabel}{y-axis label.}

\item{legend_nrow}{Number of rows for the legend to span over.}

\item{label_yaxis}{Whether or not to label the y-axis.}

\item{top_CS_only}{Only include the top 1 CS per fine-mapping method.}

\item{show_plot}{Show plot.}
}
\description{
Loci ordered by UCS size (smallest to largest).
}
\examples{
dat <- echodata::BST1
dat$Locus <- "BST1"
gg_CS <- echoannot::CS_counts_plot(merged_DT = dat)
}
\seealso{
Other summarise: 
\code{\link{CS_bin_plot}()},
\code{\link{get_SNPgroup_counts}()},
\code{\link{peak_overlap_plot}()},
\code{\link{plot_dataset_overlap}()},
\code{\link{super_summary_plot}()}
}
\concept{summarise}
