% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(
  fname,
  repo = "RajLabMSSM/echoannot",
  save_dir = tools::R_user_dir("echoannot", which = "cache"),
  overwrite = FALSE
)
}
\description{
Download remote resources stored on GitHub Releases via \pkg{piggyback}.
}
\keyword{internal}
