% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_supplementary_files.R
\name{get_geo_supplementary_files}
\alias{get_geo_supplementary_files}
\title{Get GEO supplementary files.}
\usage{
get_geo_supplementary_files(
  gsm,
  regex_queries = list(narrowPeak = "narrowpeak", broadPeak = "broadpeak", genericPeak
    = "peak", bedGraph = "bedgraph|graph.gz|bdg.gz", bigWig = "bigwig|bw$"),
  verbose = TRUE
)
}
\arguments{
\item{gsm}{GEO GSM id.}

\item{regex_queries}{Named list of queries.
Matches will be case-insensitive.}

\item{verbose}{Print messages.}
}
\description{
Get links to any GEO supplementary files matches one or more of the 
\code{regex_queries}.
}
\keyword{internal}
