% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SNPgroup_counts.R
\name{get_SNPgroup_counts}
\alias{get_SNPgroup_counts}
\title{Tally locus-specific SNP group sizes}
\usage{
get_SNPgroup_counts(merged_DT, grouping_vars = "Locus")
}
\description{
Tally locus-specific SNP group sizes
}
\examples{
data("merged_DT")
snp_groups <- get_SNPgroup_counts(merged_DT = echodata::get_Nalls2019_merged())
}
\seealso{
Other summarise: 
\code{\link{CS_bin_plot}()},
\code{\link{CS_counts_plot}()},
\code{\link{get_CS_bins}()},
\code{\link{get_CS_counts}()},
\code{\link{peak_overlap_plot}()},
\code{\link{plot_dataset_overlap}()},
\code{\link{results_report}()},
\code{\link{super_summary_plot}()}
}
\concept{summarise}
