% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOTIFBREAKR.R
\name{MOTIFBREAKR}
\alias{MOTIFBREAKR}
\title{Run \pkg{motifbreakR}}
\source{
\href{https://pubmed.ncbi.nlm.nih.gov/26272984/}{Publication}
\href{https://github.com/Simon-Coetzee/MotifBreakR}{GitHub}
\href{http://simon-coetzee.github.io/motifBreakR}{Vingette}
}
\usage{
MOTIFBREAKR(
  rsid_list,
  results_dir = file.path(tempdir(), "results"),
  pwmList = NULL,
  pwmList_max = NULL,
  genome_build = NULL,
  organism = "Hsapiens",
  threshold = 0.85,
  show.neutral = FALSE,
  method = "default",
  calculate_pvals = TRUE,
  force_new = FALSE,
  background = c(A = 0.25, C = 0.25, G = 0.25, T = 0.25),
  granularity = NULL,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{rsid_list}{RSIDs of SNPs to test for motif disruption 
between the reference and alternative alleles..}

\item{results_dir}{Directory where results should be saved 
as a file named:
 \emph{<results_dir>/_genome_wide/motifbreakR/motifbreakR_results.rds}.
If \code{NULL}, results will not be saved to disk.}

\item{pwmList}{An object of class \code{MotifList} containing the motifs that
you wish to interrogate}

\item{pwmList_max}{Limit the maximum number of PWM datasets tested 
(e.g. \code{10}). If \code{NULL}, no limit it set.}

\item{genome_build}{Genome build to use.}

\item{organism}{Only include datasets in the \code{pwmList} 
performed in a particular organism.}

\item{threshold}{Numeric; the maximum p-value for a match to be called or a minimum score threshold}

\item{show.neutral}{Logical; include neutral changes in the output}

\item{method}{Character; one of \code{default}, \code{log}, \code{ic}, or \code{notrans}; see
details.}

\item{calculate_pvals}{Calculate p-values for all SNPs tested.
\emph{WARNING:} May take a long time if many SNPs and/or PWM are selected.}

\item{force_new}{If results of the same name already exist, 
overwrite them with new analyses (\code{TRUE}). 
Otherwise, import the existing results and skip the analyses
 (default: \code{FALSE}).}

\item{background}{Numeric Vector; the background probabilities of the nucleotides}

\item{granularity}{Numeric Vector; the granularity to which to round the PWM,
larger values compromise full accuracy for speed of calculation. A value of
\code{NULL} does no rounding.}

\item{nThread}{Number of threads to parallelize analyses across.}

\item{verbose}{Logical; if running serially, show verbose messages}
}
\value{
Motif disruption predictions in 
 \link[GenomicRanges]{GRanges} format.
}
\description{
\pkg{motifbreakR} is a package to predict how much a SNP will disrupt
a transcription factor binding motif (if it falls within one).
\emph{Notes:}\cr
\itemize{
\item{\pkg{BSgenome}}{Users must manually run \code{library(BSgenome)} 
before running any \pkg{motifbreakR} functions 
to successfully use this tool.
}
\item{\code{threshold=}}{
 If \code{filterp=TRUE}, this argument indicates the p-value threshold.
 If \code{filterp=FALSE}, this argument instead indicates the pct threshold.
 }
}
}
\examples{
library(BSgenome) ## <-- IMPORTANT!
#### Example fine-mapping results ####
merged_DT <- echodata::get_Nalls2019_merged()
#### Run motif analyses ####
mb_res <- MOTIFBREAKR(rsid_list = c("rs11175620"),
                      # limit the number of datasets tested 
                      # for demonstration purposes only
                      pwmList_max = 5,
                      calculate_pvals = FALSE)
}
\seealso{
Other motifbreakR: 
\code{\link{MOTIFBREAKR_filter_by_metadata}()}
}
\concept{motifbreakR}
