% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOTT2019_epigenomic_histograms.R
\name{NOTT2019_epigenomic_histograms}
\alias{NOTT2019_epigenomic_histograms}
\title{Plot brain cell-specific epigenomic data}
\source{
\href{https://science.sciencemag.org/content/366/6469/1134}{Nott et al. (2019)}
\url{https://genome.ucsc.edu/cgi-bin/hgTracks?db=hg19&lastVirtModeType=default&lastVirtModeExtraState=&virtModeType=default&virtMode=0&nonVirtPosition=&position=chr2:127770344-127983251&hgsid=778249165_ySowqECRKNxURRn6bafH0yewAiuf}
}
\usage{
NOTT2019_epigenomic_histograms(
  dat,
  locus_dir = tempdir(),
  show_plot = TRUE,
  save_plot = FALSE,
  full_data = TRUE,
  bigwig_dir = NULL,
  return_assay_track = FALSE,
  binwidth = 200,
  density_adjust = 0.2,
  plot.zoom = "1x",
  strip.text.y.angle = 90,
  xtext = TRUE,
  geom = "density",
  plot_formula = "Cell_type ~.",
  fill_var = "Assay",
  genomic_units = "Mb",
  as_ggplot = TRUE,
  dpi = 300,
  height = 15,
  width = 8,
  nThread = 1,
  save_annot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Fine-mapping results data from \link[echolocatoR]{finemap_loci}.}

\item{dpi}{dpi to use for raster graphics}

\item{height}{height (defaults to the height of current
  plotting window)}

\item{width}{width (defaults to the width of current
  plotting window)}
}
\description{
Brain cell-specific epigenomic data from Nott et al. (2019).
}
\examples{
nott2019_track <- echoannot::NOTT2019_epigenomic_histograms(
    dat = echodata::BST1)
}
\seealso{
Other NOTT2019: 
\code{\link{NOTT2019_bigwig_metadata}},
\code{\link{NOTT2019_get_epigenomic_peaks}()},
\code{\link{NOTT2019_get_interactions}()},
\code{\link{NOTT2019_get_interactome}()},
\code{\link{NOTT2019_get_promoter_celltypes}()},
\code{\link{NOTT2019_get_promoter_interactome_data}()},
\code{\link{NOTT2019_get_regulatory_regions}()},
\code{\link{NOTT2019_plac_seq_plot}()},
\code{\link{NOTT2019_superenhancers}()},
\code{\link{get_NOTT2019_interactome}()},
\code{\link{get_NOTT2019_superenhancer_interactome}()}
}
\concept{NOTT2019}
