% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOTT2019_get_regulatory_regions.R
\name{NOTT2019_get_regulatory_regions}
\alias{NOTT2019_get_regulatory_regions}
\title{Plot brain cell-specific epigenomic data}
\source{
\href{https://science.sciencemag.org/content/366/6469/1134}{Nott et al. (2019)}
\url{https://genome.ucsc.edu/cgi-bin/hgTracks?db=hg19&lastVirtModeType=default&lastVirtModeExtraState=&virtModeType=default&virtMode=0&nonVirtPosition=&position=chr2:127770344-127983251&hgsid=778249165_ySowqECRKNxURRn6bafH0yewAiuf}
}
\usage{
NOTT2019_get_regulatory_regions(
  as.granges = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\description{
Plot brain cell-specific epigenomic data
}
\seealso{
Other NOTT2019: 
\code{\link{NOTT2019_bigwig_metadata}},
\code{\link{NOTT2019_epigenomic_histograms}()},
\code{\link{NOTT2019_get_epigenomic_peaks}()},
\code{\link{NOTT2019_get_interactions}()},
\code{\link{NOTT2019_get_interactome}()},
\code{\link{NOTT2019_get_promoter_celltypes}()},
\code{\link{NOTT2019_get_promoter_interactome_data}()},
\code{\link{NOTT2019_plac_seq_plot}()},
\code{\link{NOTT2019_superenhancers}()},
\code{\link{get_NOTT2019_interactome}()},
\code{\link{get_NOTT2019_superenhancer_interactome}()}
}
\concept{NOTT2019}
\keyword{internal}
