% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOTT2019_plac_seq_plot.R
\name{NOTT2019_plac_seq_plot}
\alias{NOTT2019_plac_seq_plot}
\title{Plot brain cell-specific interactome data}
\source{
\href{https://doi.org/10.1126/science.aay0793}{Nott et al. (2019)}
}
\usage{
NOTT2019_plac_seq_plot(
  dat = NULL,
  locus_dir = NULL,
  title = NULL,
  print_plot = TRUE,
  save_plot = TRUE,
  return_interaction_track = FALSE,
  x_limits = NULL,
  zoom_window = NULL,
  index_SNP = NULL,
  genomic_units = "POS",
  color_dict = c(enhancers = "springgreen2", promoters = "purple", anchors = "black"),
  highlight_plac = TRUE,
  show_regulatory_rects = TRUE,
  show_anchors = TRUE,
  strip.text.y.angle = 0,
  xtext = TRUE,
  save_annot = FALSE,
  point_size = 2,
  height = 7,
  width = 7,
  dpi = 300,
  return_as = "Tracks",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Fine-mapping results data from \link[echolocatoR]{finemap_loci}.}

\item{locus_dir}{Locus-specific directory.}

\item{title}{all title elements: plot, axes, legends (\code{\link[ggplot2:element]{element_text()}};
inherits from \code{text})}

\item{print_plot}{Print plot.}

\item{save_plot}{Whether to save the plot.}

\item{return_interaction_track}{Return only the interaction track 
(before completing the plot and showing it).}

\item{x_limits}{x-axis limits to be applied to all plots
(useful when trying to keep a common coordinate system).}

\item{zoom_window}{Zoom window.}

\item{index_SNP}{Index/lead SNP RSID.}

\item{genomic_units}{Which genomic units to return window limits in.}

\item{color_dict}{Named list of colors for each regulatory element.}

\item{highlight_plac}{Whether to scale opacity of PLAC-seq interactions 
(arches) such that interactions with anchors containing Consensus SNPs 
will be colored darker (Default: \code{TRUE}).
If \code{FALSE}, will instead apply the same opacity level
 to all interactions.}

\item{show_regulatory_rects}{Show enhancers/promoters as rectangles.}

\item{show_anchors}{Show PLAC-seq anchors.}

\item{strip.text.y.angle}{Angle of the y-axis facet labels.}

\item{xtext}{Whether to include x-axis title and text.}

\item{save_annot}{Save the queried subset of bigwig annotations.}

\item{point_size}{Point size of each SNP in the GWAS/fine-mapping plots.}

\item{height}{height (defaults to the height of current
  plotting window)}

\item{width}{width (defaults to the width of current
  plotting window)}

\item{dpi}{dpi to use for raster graphics}

\item{return_as}{Plot class to convert \code{plot_list} to:
\itemize{
\item{"ggplot"}{\link[ggplot2]{ggplot}}
\item{"ggbio"}{\link[ggbio]{ggbio}}
\item{"patchwork"}{\pkg{patchwork}}
\item{"Tracks"}{\link[ggbio]{tracks}}
\item{NULL}{Return original object.}
}}

\item{nThread}{Number of threads to parallelise downloads across.}

\item{verbose}{Print messages.}
}
\description{
Plot brain cell-specific interactome data from
 \href{https://doi.org/10.1126/science.aay0793}{Nott et al. (2019)}.
}
\examples{
trks_plus_lines <- echoannot::NOTT2019_plac_seq_plot(dat = echodata::BST1) 
}
\seealso{
Other NOTT2019: 
\code{\link{NOTT2019_bigwig_metadata}},
\code{\link{NOTT2019_epigenomic_histograms}()},
\code{\link{NOTT2019_get_epigenomic_peaks}()},
\code{\link{NOTT2019_get_interactions}()},
\code{\link{NOTT2019_get_interactome}()},
\code{\link{NOTT2019_get_promoter_celltypes}()},
\code{\link{NOTT2019_get_promoter_interactome_data}()},
\code{\link{NOTT2019_get_regulatory_regions}()},
\code{\link{NOTT2019_superenhancers}()},
\code{\link{get_NOTT2019_interactome}()},
\code{\link{get_NOTT2019_superenhancer_interactome}()}
}
\concept{NOTT2019}
