% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missense.R
\name{plot_missense}
\alias{plot_missense}
\title{Plot any missense variants}
\usage{
plot_missense(
  merged_DT,
  snp_filter = "Support>0",
  label_yaxis = FALSE,
  x_label = "UCS missense\\nmutations",
  show.legend = TRUE,
  show_numbers = FALSE,
  show_plot = TRUE
)
}
\description{
Plot any missense variants
}
\examples{
\dontrun{
merged_DT <- echodata::get_Nalls2019_merged()
gg_missense <- plot_missense(
    merged_DT = merged_DT,
    snp_filter = "Support>0"
)
gg_missense <- plot_missense(
    merged_DT = merged_DT,
    snp_filter = "Consensus_SNP==TRUE"
)
}
}
\seealso{
Other annotate: 
\code{\link{annotate_missense}()},
\code{\link{biomart_snp_info}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{haplor_epigenetics_enrichment}()},
\code{\link{haplor_epigenetics_summary}()},
\code{\link{haplor_haploreg}()},
\code{\link{haplor_regulomedb}()},
\code{\link{snps_by_mutation_type}()}
}
\concept{annotate}
