% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP_query.R
\name{ROADMAP_query}
\alias{ROADMAP_query}
\title{Query Roadmap by genomic coordinates}
\usage{
ROADMAP_query(
  query_dat,
  results_path = file.path(tempdir(), "Roadmap"),
  keyword_query = NULL,
  limit_files = NULL,
  chrom_states = NULL,
  remove_tmps = TRUE,
  return_paths = FALSE,
  conda_env = "echoR_mini",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{query_dat}{Variant-level summary statistics.}

\item{results_path}{Where to store query results.}

\item{keyword_query}{Search all columns in the Roadmap annotations metadata
and only query annotations that contain your keywords.
Can provide multiple keywords in list form:
\code{c("placenta","liver","monocytes")}.}

\item{limit_files}{Limit the number of annotation files queried
(for faster testing).}

\item{chrom_states}{Filter results by chromatin states.}

\item{return_paths}{Return list of paths instead of a
\link[GenomicRanges]{GRangesList}.}

\item{conda_env}{Conda environment to search for tabix in.}

\item{nThread}{Number of threads to parallelise queries over.}

\item{verbose}{Print messages.}
}
\description{
Query Roadmap by genomic coordinates
}
\examples{
query_dat <- echodata::BST1
grl <- ROADMAP_query(
    query_dat = query_dat,
    limit_files = 1,
    keyword_query = "placenta")
}
\seealso{
Other ROADMAP: 
\code{\link{ROADMAP_construct_reference}()},
\code{\link{ROADMAP_merge_and_process_grl}()},
\code{\link{ROADMAP_tabix}()}
}
\concept{ROADMAP}
