% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP_query.R
\name{ROADMAP_query}
\alias{ROADMAP_query}
\title{Query Roadmap by genomic coordinates}
\usage{
ROADMAP_query(
  results_path = file.path(tempdir(), "Roadmap"),
  gr.snp,
  keyword_query = NULL,
  limit_files = NULL,
  remove_tmps = TRUE,
  conda_env = "echoR_mini",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{results_path}{Where to store query results.}

\item{gr.snp}{\code{\link[GenomicRanges]{GRanges}} object of
SNPs to query Roadmap with.}

\item{keyword_query}{Search all columns in the Roadmap annotations metadata
and only query annotations that contain your keywords.
Can provide multiple keywords in list form:
\code{c("placenta","liver","monocytes")}}

\item{limit_files}{Limit the number of annotation files queried
(for faster testing).}

\item{remove_tmps}{Remove temporary files (e.g. "*.tbi").}

\item{conda_env}{Conda environment to search for tabix in.}

\item{nThread}{Number of threads to parallelise queries over.}

\item{verbose}{Print messages.}
}
\description{
Query Roadmap by genomic coordinates
}
\examples{
grl.roadmap <- echoannot::ROADMAP_query(
    gr.snp = echodata::BST1,
    keyword_query = "placenta"
)
}
\seealso{
Other ROADMAP: 
\code{\link{ROADMAP_construct_reference}()},
\code{\link{ROADMAP_merge_and_process_grl}()},
\code{\link{ROADMAP_tabix}()}
}
\concept{ROADMAP}
