% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP_tabix.R
\name{ROADMAP_tabix}
\alias{ROADMAP_tabix}
\title{Query Roadmap API}
\source{
\code{
query_dat <- echodata::BST1
dat <- ROADMAP_tabix(
    query_dat = query_dat,
    eid = "E099")
}
}
\usage{
ROADMAP_tabix(
  eid,
  query_dat,
  save_dir = tempdir(),
  conda_env = "echoR_mini",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{eid}{Roadmap annotation ID}

\item{query_dat}{Variant-level summary statistics.}

\item{verbose}{Print messages.}

\item{results_path}{Where to store query results.}

\item{as_granges}{Whether to return query
as a \code{data.frame} or \link[GenomicRanges]{GRanges}.}
}
\description{
Query Roadmap epigenomic annotations (chromatin marks)
using a range of genomic coordinates.
}
\details{
\href{https://egg2.wustl.edu/roadmap/data/byFileType/chromhmmSegmentations/ChmmModels/coreMarks/jointModel/final/}{
ROADMAP file locations.}
}
\seealso{
Other ROADMAP: 
\code{\link{ROADMAP_construct_reference}()},
\code{\link{ROADMAP_merge_and_process_grl}()},
\code{\link{ROADMAP_query}()}
}
\concept{ROADMAP}
\keyword{internal}
