% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_summary_plot.R
\name{super_summary_plot}
\alias{super_summary_plot}
\title{Merge all summary plots into one super plot}
\usage{
super_summary_plot(
  merged_DT,
  snp_filter = "Consensus_SNP==TRUE",
  coloc_results = NULL,
  plot_missense = TRUE,
  show_plot = TRUE,
  save_plot = FALSE,
  height = 15,
  width = 13,
  dpi = 300
)
}
\arguments{
\item{merged_DT}{Merged fine-mapping results data from
\link[echolocatoR]{finemap_loci}.}

\item{snp_filter}{Filter to use apply to SNPs before plotting.}

\item{coloc_results}{Colocalization results from 
\href{https://github.com/RajLabMSSM/catalogueR}{catalogueR}.}

\item{plot_missense}{Whether to include the missense mutations plot.
\emph{Warning:} Can take a lot time to query Biomart.}

\item{show_plot}{Show plot.}

\item{save_plot}{Save plot.}

\item{height}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{width}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}
}
\description{
Merge all summary plots into one super plot
}
\examples{
\dontrun{
merged_DT <- echodata::get_Nalls2019_merged() 
super_plot <- echoannot::super_summary_plot(merged_DT = merged_DT,
                                            plot_missense = FALSE)
}
}
\seealso{
Other summarise: 
\code{\link{CS_bin_plot}()},
\code{\link{CS_counts_plot}()},
\code{\link{get_SNPgroup_counts}()},
\code{\link{peak_overlap_plot}()},
\code{\link{plot_dataset_overlap}()},
\code{\link{results_report}()}
}
\concept{summarise}
