% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_enrichment.R
\name{test_enrichment}
\alias{test_enrichment}
\title{Test enrichment}
\usage{
test_enrichment(
  grlist1,
  grlist2,
  ntimes = 50,
  genome = "hg19",
  alternative = "auto",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{grlist1}{First list of \link[GenomicRanges]{GRanges} objects.}

\item{grlist2}{Second list of \link[GenomicRanges]{GRanges} objects.}

\item{ntimes}{number of permutations}

\item{genome}{The reference genome to use. A valid genome object. 
Either a GenomicRanges or data.frame containing one region per whole 
chromosome or a character uniquely identifying a genome in 
BSgenome (e.g. "hg19", "mm10" but not "hg"). 
Internally it uses getGenomeAndMask.}

\item{alternative}{the alternative hypothesis must be one of \code{"greater"}, \code{"less"} or \code{"auto"}. If \code{"auto"}, the alternative will be decided depending on the data.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments passed to \link[regioneR]{overlapPermTest}.}
}
\description{
Conduct permutation enrichment tests between all combinations of 
two named lists,
each containing one or more \link[GenomicRanges]{GRanges} objects.
Permutation tests are run using \link[regioneR]{overlapPermTest}.
}
